% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.R
\name{add_strip_pal}
\alias{add_strip_pal}
\title{Add a palette to the strip of a plot}
\usage{
add_strip_pal(plot, pal)
}
\arguments{
\item{plot}{The plot to add the palette to.}

\item{pal}{The palette to add to the strip.}
}
\value{
The plot with the palette added to the strip.
}
\description{
This function adds a palette to the strip of a plot.
}
