% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-plotFunctions.R
\name{addLegendServer}
\alias{addLegendServer}
\title{Add a legend to an existing plot}
\usage{
addLegendServer(id, position, title, legendEntries, fillColor)
}
\arguments{
\item{id}{module id}

\item{position}{legend position, e.g. "bottomright", "bottom", "bottomleft", "left", "topleft", "top", "topright", "right" and "center"}

\item{title}{legend title}

\item{legendEntries}{names of legend entries}

\item{fillColor}{color for the legend symbols}
}
\description{
Add a legend to an existing plot
}
