% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-plotFunctions.R
\name{addLegendUI}
\alias{addLegendUI}
\title{selectInput to specify legend position}
\usage{
addLegendUI(
  id,
  label = "Legend Position",
  choices = c("none", "topleft", "topright", "bottomright", "bottomleft")
)
}
\arguments{
\item{id}{module id}

\item{label}{label of button}

\item{choices}{choices e.g. "bottomright", "bottom", "bottomleft", "left", "topleft", "top", "topright", "right" and "center"}
}
\description{
selectInput to specify legend position
}
