% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-addPoints.R
\name{addPointsUI}
\alias{addPointsUI}
\alias{addPoints}
\title{Add points module}
\usage{
addPointsUI(id)

addPoints(input, output, session)
}
\arguments{
\item{id}{id of module}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\description{
UI function to define point values for a point to be added to the plot

Backend for add points module
}
