% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_functions.R
\name{arrange_ggplots}
\alias{arrange_ggplots}
\title{Arrange a list of plots}
\usage{
arrange_ggplots(
  plotlist,
  title = NULL,
  remove_subtitles = F,
  new_subtitles = NULL,
  legend = "right",
  title_size = 24,
  subtitle_size = 18,
  remove_individual_axes_titles = T,
  xaxis_title = NULL,
  yaxis_title = NULL,
  axis_title_size = 18,
  nrow = length(plotlist),
  ncol = length(plotlist)/nrow,
  align = "hv"
)
}
\arguments{
\item{plotlist}{A list of ggplot objects}

\item{title}{Title for arranged plots}

\item{remove_subtitles}{A logical value indicating whether or not to remove titles from individual plots. Default is FALSE.}

\item{new_subtitles}{An optional vector with new titles to add to the plots.}

\item{legend}{Character specifying where to put legend. Default is to the right. Set to "none" to remove.}

\item{title_size}{Font size for title. Default is 24}

\item{subtitle_size}{Size of the subtitles. Default is 18}

\item{remove_individual_axes_titles}{A logical value indicating whether to remove individual axes titles. Default is TRUE.}

\item{xaxis_title}{Title to use for x-axis.}

\item{yaxis_title}{Title to use for y-axis}

\item{axis_title_size}{Size for the axes titles. Default is 18.}

\item{nrow}{Number of rows for combined plot.}

\item{ncol}{Number of columns for combined plot}

\item{align}{How to align plots. Same as for ggarrange(). Default is to align vertically and horizontally ("hv").}
}
\value{
A ggarrange object
}
\description{
Arrange a list of plots
}
