% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_extra.R
\name{barlabel}
\alias{barlabel}
\title{Bar plot labels}
\usage{
barlabel(
  x,
  labels = x,
  at = barplot(x, plot = FALSE, horiz = horiz, beside = beside),
  horiz = FALSE,
  beside = FALSE,
  hadj = 0,
  vadj = 0,
  position = c("middle", "top", "bottom"),
  plot = TRUE,
  ...
)
}
\arguments{
\item{x}{a vector or matrix of bar heights, typically the same object used
in \code{barplot(x)}}

\item{labels}{labels for each bar segment}

\item{at}{the position of the midpoint of each bar}

\item{horiz}{logical; plot vertical (default) or horizontal bars}

\item{hadj, vadj}{optional horizontal and vertical adjustments to the
calculated coordinates}

\item{position}{the position for labels relative to each bar segment, one of
\code{"middle"} (default), \code{"top"}, or \code{"bottom"}}

\item{plot}{logical; if \code{TRUE}, labels are drawn; otherwise, calculated
coordinates are returned invisibly}

\item{...}{additional graphical parameters passed to \code{\link{text}} or
further to \code{\link{par}}}
}
\value{
A list containing the x- and y-coordinates at which \code{labels} were drawn.
}
\description{
Add labels to a \code{\link{barplot}}.
}
\examples{
x <- table(mtcars$gear)
barplot(x)
barlabel(x, position = 'top', pos = 3L)

x <- table(mtcars$gear, mtcars$vs)
barplot(x)
barlabel(x, labels = sprintf('n = \%s (\%.1f\%\%)', x, x / colSums(x) * 109),
         col = cm.colors(nrow(x)))

bp <- barplot(x, beside = TRUE)
barlabel(x, beside = TRUE)

}
