% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2.R
\name{barprop}
\alias{barprop}
\title{barprop}
\usage{
barprop(
  height,
  col = NULL,
  labels = height,
  col.labels = par("col.lab"),
  horiz = TRUE,
  gap = NULL,
  min.prop = Inf,
  axis.pos = NULL,
  xlab = NULL,
  ylab = NULL,
  ...
)
}
\arguments{
\item{height}{a matrix (or vector) of proportions}

\item{col}{colors for each group, i.e., \code{nrow(height)} colors}

\item{labels}{a matrix or vector of labels for each bar plot segment; if
\code{FALSE}, no labels are drawn}

\item{col.labels}{a matrix or vector of colors for each label, i.e.,
\code{length(height)} colors}

\item{horiz}{logical; if \code{TRUE} (default), bars are drawn horizontally
with the first at the bottom; otherwise, bars are drawn vertically with
the first to the left}

\item{gap}{the space between each group as a single value or vector in units
of \code{height}, recycled as needed}

\item{min.prop}{the minimum proportion controlling which labels are drawn
in the center of each bar; for any \code{height > min.prop}, labels are
drawn to the right of the bar (or to the top if \code{horiz = FALSE})}

\item{axis.pos}{a vector of length 2 giving the position of the x- and
y-axes; note that \code{axis.pos[1L]} can only be 1 or 3 and
\code{axis.pos[2L]} can only be 2 or 4}

\item{xlab, ylab}{the x- and y-axis labels, default is
\code{names(dimnames(height))}}

\item{...}{additional graphical parameters passed to \code{\link{par}}

\code{labels} (i.e., the labels for each bar segment) may be customized
using \code{col.labels}; \code{cex.lab}, and/or \code{font.lab}

the axis labels (i.e., the labels for each group) may be customized
using \code{col.axis}; \code{cex.axis}, and/or \code{font.axis}

the main labels (i.e., \code{xlab} and \code{ylab}) may be customized using
\code{col.main}; \code{cex.main}, and/or \code{font.main} if given}
}
\description{
A \code{\link{barplot}} for proportions. This function is intended to show
a matrix of proportions as a stacked bar plot but each group will have a
common starting height. Therefore, it is recommended that each column sums
to 100% (or 1 for proportions); however, this is not required.
}
\examples{
x <- table(Gear = mtcars$gear, Cylinder = mtcars$cyl)
dimnames(x) <- lapply(dimnames(x), function(y) rawr::num2char(as.numeric(y)))
p <- round(prop.table(x, 2) * 100)
barprop(p)

barprop(x, horiz = FALSE)
## compare
barplot(x)

## on vectors (or single-column matrices)
barprop(p[, 1])
barprop(p[, 1, drop = FALSE])

## options/aesthetics
barprop(p, min.prop = 0, col.labels = ifelse(row(p) == 1, 'white', 'black'))
barprop(p, gap = 10)
barprop(p, min.prop = 0, gap = c(10, 0))
barprop(x, min.prop = 0, labels = sprintf('\%s\%\%', x), cex.lab = 1.5,
  font.axis = 2, cex.axis = 1.5)

}
