% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{bibar}
\alias{bibar}
\title{Bi-barplot}
\usage{
bibar(x, left = NULL, right = NULL, sleft = NULL, sright = NULL,
  col = NULL, xlim = NULL, ylim = NULL, axes = TRUE)
}
\arguments{
\item{x}{a table- or matrix-like object}

\item{left, right}{a vector of integers giving the column indices of
\code{x} to draw on the left and right sides \emph{from} the origin
\emph{to} either side}

\item{sleft, sright}{as above but bars drawn \emph{from} the left or
right side \emph{to} the origin}

\item{col}{a vector of colors for each column}

\item{xlim, ylim}{x- and y-axis limits}

\item{axes}{logical; if \code{TRUE}, the figure is framed and the x- and
y-axes are drawn}
}
\description{
Plot a bi-directional bar plot with optional side bars.
}
\examples{
set.seed(1)
x <- datasets::ability.cov$cov
x <- x[sample(seq.int(nrow(x)), 20, TRUE), ]

bibar(x, left = 1:3, right = 4:6, xlim = c(-250, 250))

palette(c('grey90', 'cornflowerblue', 'blue', 'tomato', 'tomato3'))
bibar(x, left = 2:3, right = 4:5, sleft = 1, sright = 6)
legend('topleft', inset = c(0, -0.2), xpd = NA, fill = 2:3,
       legend = colnames(x)[2:3], horiz = TRUE, bty = 'n')
legend('topright', inset = c(0, -0.2), xpd = NA, fill = 4:5,
       legend = colnames(x)[4:5], horiz = TRUE, bty = 'n')
palette('default')

}
\seealso{
\code{\link{prettybars2}}
}
