% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnw.palette.R
\name{bnw.palette}
\alias{bnw.palette}
\title{Convert colors to black and white}
\usage{
bnw.palette(color)
}
\arguments{
\item{color}{}
}
\description{
Converts a vector of HEX values to their respective black and white variants, also in HEX.
}
