% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_misc.R
\name{boxline}
\alias{boxline}
\title{Boxline}
\usage{
boxline(x, probs = c(0.9, 0.95, 0.99), at = seq_along(x), col = 2L,
  alpha = NULL, add = FALSE, ...)
}
\arguments{
\item{x}{a list of vectors}

\item{probs}{a vector of probabilities; see \code{\link{quantile}}}

\item{at}{a vector of x-coodinates for each vector of \code{x}}

\item{col}{a vector of colors for each \code{prob}, recycled as needed}

\item{alpha}{optional vector of opacity for each \code{prob}; see
\code{\link{adjustcolor}}}

\item{add}{logical; if \code{TRUE}, adds to existing plot}

\item{...}{additional arguments passed to \code{\link{boxplot}} or further
to \code{\link{par}}}
}
\description{
An alternative to \code{\link{boxplot}} with lines connecting quantiles of
each box.
}
\examples{
set.seed(1)
x <- lapply(0:10, function(x) rnorm(10, x / 2, sd = 0.5))
boxplot(x)
boxline(x, add = TRUE)

boxline(x, col = c('red', 'orange', 'yellow'), alpha = 0.5, las = 1L)
boxplot(x, add = TRUE, axes = FALSE)

}
