% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{bump}
\alias{bump}
\title{bump chart}
\usage{
bump(mat, adj = 0.5, ...)
}
\arguments{
\item{mat}{an \code{n x t} matrix with \code{n} observations and \code{t}
timepoints}

\item{adj}{label position adjustment}

\item{...}{additional graphical parameters passed to \code{par}}
}
\description{
A bump chart
}
\examples{
mat <- replicate(5, sample(1:10))
dimnames(mat) <- list(rownames(mtcars)[1:nrow(mat)], paste0('time', 1:ncol(mat)))
bump(mat, mar = c(2,4,2,9), adj = .1, bg = 'cornsilk')

}

