% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_extra.R
\name{caxis}
\alias{caxis}
\title{Add a color axis}
\usage{
caxis(side, prop, col = NULL, lwd = 3, ...)
}
\arguments{
\item{side}{an integer specifying which side of the plot the axis is to be
drawn on: 1=below, 2=left, 3=above, and 4=right}

\item{prop}{the proportion of each section to be scaled to the axis width}

\item{col}{a vector of colors equal to the length of \code{prop}}

\item{lwd}{line width for the axis}

\item{...}{additional parameters passed to \code{\link{segments}}}
}
\value{
A list containing the start and end positions for each \code{prop} scaled
to the axis width.
}
\description{
Add a color-coded axis.
}
\examples{
x <- runif(100)
y <- runif(100)

plot(x, y)
caxis(1, 1:4, col = 2:5)
at <- caxis(3, 1:4, col = 2:5)
text(at$end, par('usr')[4], paste('group', 1:4),
     col = 2:5, xpd = NA, adj = c(1, -1))

}
