% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-multiplePredictions.R
\name{checkReq}
\alias{checkReq}
\title{Pop up a Warning if Req is missing.}
\usage{
checkReq(x, label)
}
\arguments{
\item{x}{reactive object to be checked with req()}

\item{label}{(character) Warning that pops up if x is missing}
}
\description{
Pop up a Warning if Req is missing.
}
