% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_extra.R
\name{click_shape}
\alias{click_shape}
\title{Add shapes interactively in base \code{R} graphics}
\usage{
click_shape(
  shape = c("circle", "arrow", "rect", "cyl", "line", "poly"),
  corners = 3L,
  ...
)
}
\arguments{
\item{shape}{type of shape; choices are \code{'box'}, \code{'arrow'},
\code{'line'}, \code{'poly'}, \code{'circle'}, and \code{'cyl'}}

\item{corners}{number of corners to draw if \code{shape = 'poly'}}

\item{...}{additional arguments or graphical parameters passed to the
shape functions}
}
\description{
Add shapes anywhere in a plot (including margins) with mouse click(s).
}
\examples{
\dontrun{
op <- par(xpd = NA)
plot.new()
plot.window(0:1, 0:1, asp = 1)
click_shape('line') # a line segment
click_shape('arrow', col = 'blue', code = 2, lwd = 2, length = .15)
click_shape('rect', border = 'purple', col = 'pink', lwd = 2)
click_shape('rect', col = NULL, border = 'purple', lwd = 2)
click_shape('line', col = 'orange', lty = 3, lwd = 3)
click_shape('poly', corners = 5, border = 'green', col = 'orange')
click_shape('poly', corners = 3, border = 'red', col = 'yellow', lty = 1)
click_shape('cyl', col = 'orange')
click_shape('circle', col = 'orange', border = 'black', lty = 3, lwd = 3)
par(op)
}

}
\seealso{
\code{\link{click_text}}, \code{\link{rect}},
\code{\link{arrows}}, \code{\link{rect}}, \code{\link{rect}},
\code{\link{segments}}, \code{\link{polygon}},
\code{\link[plotrix]{draw.circle}}, \code{\link[plotrix]{cylindrect}}
}
