% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_extra.R
\name{color_bar}
\alias{color_bar}
\title{Color legend}
\usage{
color_bar(colors, x = NULL, y = x, labels = NULL, at.x = par("usr")[2L],
  at.y = par("usr")[3L], cex.x = 1, cex.y = 1, ...)
}
\arguments{
\item{x}{numeric vector of data}

\item{y}{optional numeric vector of data; to color by a variable other than
\code{x}, \code{y} must be specified for \code{labels} coordinates to work 
properly; see examples}

\item{labels}{optional labels for color bar}

\item{at.x}{x-coordinate of lower-left corner of bar}

\item{at.y}{y-coordinate of lower-left corner of bar; note that the bar can 
extend ouside the plotting region, so use \code{cex.x} and \code{cex.y} to
scale the bar down (or up)}

\item{cex.x, cex.y}{scaling parameters}

\item{...}{additional graphical parameters passed to \code{\link{par}}}

\item{cols}{vector of color names (or hexadecimal) from low to high}
}
\value{
An \code{\link{invisible}} vector of colors (hexadecimal format)
used for the bar. Note that these colors may be mapped to \code{x} so that 
only one \code{color_bar} needs to be called in the \code{col} argument for
both the coloring of points and the generation of the legend; see examples.
}
\description{
Continuous color bar legend.
}
\examples{
plot.new()
color_bar(c('black','red','white','blue'), at.x = 0, at.y = 0)
text(x = 0, y = seq(0, par('usr')[4], length = 6), pos = 4, xpd = TRUE,
     labels = pretty(seq(0, par('usr')[4]), n = 6), cex = .8, offset = .75)


## calling color_bar once in the col argument will color x accordingly
## and plot the legend simultaneously

## compare
plot(mtcars$mpg, pch = 19, cex = 2, 
     col = rawr:::col_scaler(mtcars$mpg, c('yellow','red')))
plot(mtcars$mpg, pch = 19, cex = 2,
     col = color_bar(c('yellow','red'), mtcars$mpg, labels = mtcars$mpg))


## plot a color_bar legend by a variable other than x
##   use y argument to match the variable in the original plot call
##   and x as the variable to color and label

## compare
with(mtcars, {
  plot(mpg, pch = 19, cex = 2,
       main = 'color by weight (red = heavier)',
       col = rawr:::col_scaler(wt, c('yellow','red')))
})

with(mtcars, {
  plot(mpg, pch = 19, cex = 2,
       main = 'color by weight (red = heavier)',
       col = color_bar(c('yellow','red'), x = wt, y = mpg, labels = wt))
})

}
