% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{color_palettes}
\alias{color_palettes}
\alias{pal_godsnot_102}
\alias{pal_zeileis_28}
\alias{pal_royal_4}
\alias{pal_asteroid_5}
\alias{pal_darjeeling_5}
\alias{pal_zissou_5}
\alias{pal_kelly_20}
\alias{pal_greenarmytage_25}
\alias{pal_brewerplus_41}
\alias{pal_calc_11}
\alias{pal_piyg_11}
\alias{pal_brbg_11}
\alias{pal_prgn_11}
\alias{pal_blackpink_6}
\alias{pal_vega10}
\alias{pal_vega20}
\alias{pal_vega20c}
\title{Color palettes from various sources}
\format{
Character vector of 102 colors

Character vector of 28 colors

Character vector of 4 colors

Character vector of 5 colors

Character vector of 5 colors

Character vector of 5 colors

Character vector of 20 colors

Character vector of 25 colors

Character vector of 41 colors

Character vector of 11 colors

Character vector of 11 colors

Character vector of 11 colors

Character vector of 11 colors

Character vector of 6 colors

Character vector of 10 colors

Character vector of 20 colors

Character vector of 20 colors
}
\usage{
pal_godsnot_102()

pal_zeileis_28()

pal_royal_4()

pal_asteroid_5()

pal_darjeeling_5()

pal_zissou_5()

pal_kelly_20()

pal_greenarmytage_25()

pal_brewerplus_41()

pal_calc_11()

pal_piyg_11()

pal_brbg_11()

pal_prgn_11()

pal_blackpink_6()

pal_vega10()

pal_vega20()

pal_vega20c()
}
\description{
Color palettes from various sources
}
\section{Sources}{

\itemize{
\item pal_godsnot_102, pal_zeileis_28: scanpy palettes
\item pal_royal_4, pal_asteroid_5, pal_darjeeling_5, pal_zissou_5: WesAnderson palettes
\item pal_kelly_20, pal_greenarmytage_25: Colorblind-friendly palettes
\item pal_vega10, pal_vega20, pal_vega20c: Vega visualization library palettes
\item pal_piyg_11, pal_brbg_11, pal_prgn_11: ColorBrewer-inspired diverging palettes
\item pal_blackpink_6, pal_calc_11: Miscellaneous palettes
\item pal_brewerplus_41: Extended ColorBrewer palette
}
}

