% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_functions.R
\docType{data}
\name{colour_list}
\alias{colour_list}
\title{A list of colour vectors}
\format{
A named list of colour vectors
}
\usage{
colour_list
}
\description{
A list of colour vectors
}
\keyword{datasets}
