% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-downAndUploadModel.R
\name{combineDataAndModelOutputs}
\alias{combineDataAndModelOutputs}
\title{Combine Data and Model Outputs}
\usage{
combineDataAndModelOutputs(modelData, modelOutput)
}
\arguments{
\item{modelData}{list of model data objects}

\item{modelOutput}{list of model output objects}
}
\description{
Combine Data and Model Outputs
}
