% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-goodnessFit.R
\name{createGoodnessOfFitTable}
\alias{createGoodnessOfFitTable}
\title{Create Goodness Of Fit Table}
\usage{
createGoodnessOfFitTable(x, llog, edf)
}
\arguments{
\item{x}{(character) name of the model}

\item{llog}{(matrix) llog matrix of a model output}

\item{edf}{(numeric) effective degrees of freedom, i.e.: sum of the row variances of the llog matrix}
}
\description{
Create Goodness Of Fit Table
}
