% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_functions.R
\name{customize_ggplot_theme}
\alias{customize_ggplot_theme}
\title{Customize the theme of a ggplot object}
\usage{
customize_ggplot_theme(
  plot,
  base_theme = theme_bw(),
  title = NULL,
  plot_title_size = 24,
  xlab = NULL,
  ylab = NULL,
  axis_title_size = 20,
  axis_text_size = 18,
  x_labels = NULL,
  format_x_labels = F,
  x_labels_angle = 0,
  show_legend = T,
  fill_title = NULL,
  color_title = NULL,
  fill_colors = NULL,
  colors = NULL,
  fill_labels = waiver(),
  legend_title_size = 20,
  legend_text_size = 18,
  legend_key_size = 1,
  format_fill_labels = F,
  scale_x = NULL,
  scale_y = NULL,
  facet = NULL,
  facet_nrow = NULL,
  facet_ncol = NULL,
  facet_scales = "free",
  strip_text_size = 20,
  facet_labels = NULL
)
}
\arguments{
\item{plot}{A ggplot object}

\item{base_theme}{The base theme to use for the plot. Default is theme_bw().}

\item{title}{Title for the plot}

\item{plot_title_size}{Size of the plot title in pts. Default is 24.}

\item{xlab}{The x-axis title}

\item{ylab}{The y-axis title}

\item{axis_title_size}{Size of the axes titles in pts. Default is 20.}

\item{axis_text_size}{Size of the axes text in pts. Default is 18.}

\item{x_labels}{Optional labels to use for the x-axis}

\item{format_x_labels}{Logical value indicating whether to format the X axis labels by replacing "_" with a space and using title case. Default is FALSE.}

\item{x_labels_angle}{Number of degrees to rotate x-axis labels. Default is 0.}

\item{show_legend}{A logical value indicating whether to show the legend or not. Default is TRUE.}

\item{fill_title}{The title for fill in the legend}

\item{color_title}{The title for color in the legend}

\item{fill_colors}{An optional vector of colors to use for the fill.}

\item{colors}{An optional vector of colors to use for the color mapping.}

\item{fill_labels}{An optional vector of labels for the fill legend}

\item{legend_title_size}{Size of the legend titles in pts. Default is 20.}

\item{legend_text_size}{Size of the legend text in pts. Default is 18.}

\item{legend_key_size}{Size of the legend key in cm. Default is 1.}

\item{format_fill_labels}{Logical value indicating whether to format the legend labels by replacing "_" with a space and using title case. Default is FALSE.}

\item{scale_x}{A scale to use for the x-axis. If none provided, one is chosen by default.}

\item{scale_y}{A scale to use for the y-axis. If none provided, one is chosen by default.}

\item{facet}{A variable to use for faceting. Default is not to facet.}

\item{facet_nrow}{Number of rows to use for faceting.}

\item{facet_ncol}{Number of rows to use for faceting.}

\item{facet_scales}{scale for facet_wrap(). Default is "free".}

\item{strip_text_size}{Size of the facet labels. Default is 20.}

\item{facet_labels}{Labels to use for panels when facetting}
}
\value{
A ggplot object
}
\description{
Customize the theme of a ggplot object
}
