% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_stripplot.R
\name{data_stripplot}
\alias{data_stripplot}
\title{Stripplot}
\usage{
data_stripplot(formula, data, groups = NULL, cex = 0.8, jitter = 0.1, ...)
}
\arguments{
\item{formula}{a formula object of the form \code{y ~ x}.
\code{y} is a response, \code{x} is a covariate expression.}

\item{data}{a data frame.}

\item{groups}{an optional grouping variable.}

\item{cex}{relative character size of labels.}

\item{jitter}{amount of jittering, or 0 for no jitter.}

\item{...}{further arguments.}
}
\value{
A \code{\link[lattice]{trellis.object}}.
}
\description{
Produces a series of one-dimensional scatterplot.
}
\examples{
\dontrun{
data_stripplot(errors ~ mode, data = data2)
}
}
\seealso{
\code{\link[lattice]{stripplot}}
}
\author{
Nathan Lucas
}
