% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-dataSettings.R
\name{dataSettingsUI}
\alias{dataSettingsUI}
\alias{dataSettings}
\alias{modelSettingsUI}
\title{Data settings module}
\usage{
dataSettingsUI(id, label = "Data Settings")

dataSettings(input, output, session, data, plotData)

modelSettingsUI(id, label = "Model Settings")
}
\arguments{
\item{id}{id of module}

\item{label}{label of settings}

\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session}

\item{data}{(reactive) dataframe of the selected loaded file}

\item{plotData}{(reactive) list of selected saved data}
}
\description{
UI function to define settings for imported data

Backend for data settings module

UI function to define settings for the model
}
