% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.envir.tools.R
\name{delete.palette}
\alias{delete.palette}
\title{Delete palette preset}
\usage{
delete.palette(palette)
}
\arguments{
\item{palette}{palette to delete}
}
\description{
Deletes a palette preset. Palettes provided with the package cannot be deleted.
WARNING: changes to palettes are permanent and cannot be reverted.
}
\keyword{"Color}
\keyword{Palette"}
