% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-predictAdditionalData.R
\name{deriveExplanatory}
\alias{deriveExplanatory}
\title{Derive Explanatory for More Y Input}
\usage{
deriveExplanatory(
  moreMean,
  moreSD,
  plotValues,
  graphName,
  moreNSample = NULL,
  isTest = FALSE
)
}
\arguments{
\item{moreMean}{numeric mean input value}

\item{moreSD}{numeric sd for input value}

\item{plotValues}{list of plotValues from savedPlots element}

\item{graphName}{(character) name of a savedPlots element}

\item{moreNSample}{(numeric) number of samples for the input}

\item{isTest}{TRUE if used within tests}
}
\description{
Derive Explanatory for More Y Input
}
