% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-predictAdditionalData.R
\name{deriveSingleExplanatory}
\alias{deriveSingleExplanatory}
\title{Derive Explanatory from Densities of the Estimation Sample}
\usage{
deriveSingleExplanatory(yMean, xVar, predDensities, quantile = 0.95)
}
\arguments{
\item{yMean}{numeric value for which to derive the explanatory}

\item{xVar}{numeric vector of possible values for the explanatory}

\item{predDensities}{list of outputs from density()}

\item{quantile}{(numeric) value between (0,1)}
}
\description{
Derive Explanatory from Densities of the Estimation Sample
}
