% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{display_palettes}
\alias{display_palettes}
\title{Display available palettes}
\usage{
display_palettes(n = NULL)
}
\arguments{
\item{n}{Number of colors to display from each palette. Default is NULL (all colors).}
}
\value{
A ggplot2 object displaying the palettes
}
\description{
Creates a visualization of all available color palettes in the package
}
\examples{
\dontrun{
# Display all palettes
display_palettes()

# Display first 5 colors of each palette
display_palettes(n = 5)
}
}
