% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-downAndUploadModel.R
\name{extractModelOutputs}
\alias{extractModelOutputs}
\title{Extract Model Outputs}
\usage{
extractModelOutputs(models)
}
\arguments{
\item{models}{list of model objects}
}
\description{
Extract Model Outputs
}
