% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fancy.axis.R
\name{fancy.axis}
\alias{fancy.axis}
\title{Plot axis with rug}
\usage{
fancy.axis(
  side = 1,
  space.main,
  tk.main,
  number.sub = 0,
  lim,
  tk.size = -0.05,
  plot.lab = TRUE,
  tk.lab
)
}
\arguments{
\item{side}{which side to plot the axis}

\item{space.main}{spacing between main ticks}

\item{tk.main}{placement of main ticks. Overrides space.main}

\item{number.sub}{number of sub ticks between main ticks. If a vector is provided, smaller sub ticks are plotted between sub ticks}

\item{lim}{axis limits}

\item{tk.size}{size of main and sub ticks. Must be of length 1+length(number.sub), otherwise values are recycled}

\item{plot.lab}{If TRUE, plots tick labels. Otherwise, tick labels are not plotted}

\item{tk.lab}{labels for main ticks}
}
\description{
Plot axis with rug
}
