% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-predictAdditionalData.R
\name{filterCommonXRange}
\alias{filterCommonXRange}
\title{Filter or restrict the common xRange}
\usage{
filterCommonXRange(commonPredictions, xRange)
}
\arguments{
\item{commonPredictions}{(data.frame) data.frame with columns xVar, and Estimations for each selected plot}

\item{xRange}{(numeric) range for which correlation will be calculated}
}
\description{
Filter or restrict the common xRange
}
