% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-fitModel.R
\name{fitModelWrapper}
\alias{fitModelWrapper}
\title{Get plotValues}
\usage{
fitModelWrapper(plotValues, modelParameters)
}
\arguments{
\item{plotValues}{(list) list with all values required for the plot}

\item{modelParameters}{(list) list of settings for the model}
}
\description{
Get plotValues
}
