% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplots.R
\name{gboxplot.matrix}
\alias{gboxplot.matrix}
\title{Draw a gboxplot for each column (row) of a matrix}
\usage{
\method{gboxplot}{matrix}(x, use.cols = TRUE, ..., grid = TRUE,
  col.grid = "grey90", col.acc = "white")
}
\arguments{
\item{x}{a numeric matrix.}

\item{use.cols}{logical indicating if columns (by default) or rows
    (\code{use.cols = FALSE}) should be plotted.}

\item{...}{Further arguments to \code{\link{boxplot}}.}

\item{grid}{logical; if \code{TRUE}, a background grid will be drawn}

\item{col.grid}{\code{grid} color}

\item{col.acc}{\code{grid} accent color}
}
\description{
Interpreting the columns (or rows) of a matrix as different groupd, draw a 
gboxplot for each.
}
