% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplots.R
\name{gbxp}
\alias{gbxp}
\title{Draw boxplots from summaries}
\usage{
gbxp(
  z,
  notch = FALSE,
  width = NULL,
  varwidth = FALSE,
  outline = TRUE,
  notch.frac = 0.5,
  log = "",
  border = par("fg"),
  pars = NULL,
  horizontal = FALSE,
  add = FALSE,
  at = NULL,
  show.names = NULL,
  ...,
  grid = TRUE,
  col.grid = "grey90",
  col.acc = "white"
)
}
\arguments{
\item{z}{a list containing data summaries to be used in constructing
    the plots.  These are usually the result of a call to
    \code{\link[graphics]{boxplot}}, but can be generated in any fashion.}

\item{notch}{if \code{notch} is \code{TRUE}, a notch is drawn in each
    side of the boxes.  If the notches of two plots do not overlap then
    the medians are significantly different at the 5 percent level.}

\item{width}{a vector giving the relative widths of the boxes making
    up the plot.}

\item{varwidth}{if \code{varwidth} is \code{TRUE}, the boxes are drawn
    with widths proportional to the square-roots of the number of
    observations in the groups.}

\item{outline}{if \code{outline} is not true, the outliers are not
    drawn.}

\item{notch.frac}{numeric in (0,1). When \code{notch = TRUE}, the
    fraction of the box width that the notches should use.}

\item{log}{character, indicating if any axis should be drawn in
    logarithmic scale, as in \code{\link[graphics]{plot.default}}.}

\item{border}{character or numeric (vector), the color of the box
    borders.  Is recycled for multiple boxes.  Is used as default for
    the \code{boxcol}, \code{medcol}, \code{whiskcol}, \code{staplecol},
    and \code{outcol} options (see below).
  }

\item{horizontal}{logical indicating if the  boxplots should be
    horizontal; default \code{FALSE} means vertical boxes.}

\item{add}{logical, if true \emph{add} boxplot to current plot.}

\item{at}{numeric vector giving the locations where the boxplots
    should be drawn, particularly when \code{add = TRUE}; defaults to
    \code{1:n} where \code{n} is the number of boxes.}

\item{show.names}{Set to \code{TRUE} or \code{FALSE} to override the
    defaults on whether an x-axis label is printed for each group.}

\item{grid}{logical; if \code{TRUE}, a background grid will be drawn}

\item{col.grid}{\code{grid} color}

\item{col.acc}{\code{grid} accent color}
}
\description{
\code{gbxp} draws box plots based on the given summaries in \code{z}. It is
usually called from within \code{\link{gboxplot}} but can be invoked 
directly.
}
