% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplots.R
\name{gcurve}
\alias{gcurve}
\alias{gplot.function}
\title{Draw function curves}
\usage{
gcurve(
  expr,
  from = NULL,
  to = NULL,
  n = 101,
  add = FALSE,
  type = "l",
  xname = "x",
  xlab = xname,
  ylab = NULL,
  log = NULL,
  xlim = NULL,
  ...,
  grid = TRUE,
  col.grid = "grey90",
  col.acc = "white"
)

\method{gplot}{`function`}(
  x,
  y = 0,
  to = 1,
  from = y,
  xlim = NULL,
  ylab = NULL,
  grid = TRUE,
  col.grid = "grey90",
  col.acc = "white",
  ...
)
}
\arguments{
\item{expr}{The name of a function, or a \link{call} or an
    \link{expression} written as a function of \code{x} which will
    evaluate to an object of the same length as \code{x}.}

\item{from}{the range over which the function will be plotted.}

\item{to}{the range over which the function will be plotted.}

\item{n}{integer; the number of x values at which to evaluate.}

\item{add}{logical; if \code{TRUE} add to an already existing plot; if
    \code{NA} start a new plot taking the defaults for the limits and
    log-scaling of the x-axis from the previous plot.  Taken as
    \code{FALSE} (with a warning if a different value is supplied) if no
    graphics device is open.}

\item{type}{plot type: see \code{\link[graphics]{plot.default}}.}

\item{xname}{character string giving the name to be used for the x axis.}

\item{xlab}{labels and \link[graphics]{graphical parameters}
    can also be specified as arguments.  See \sQuote{Details} for the
    interpretation of the default for \code{log}.

    For the \code{"function"} method of \code{plot}, \code{\dots} can
    include any of the other arguments of \code{curve}, except
    \code{expr}.
  }

\item{ylab}{labels and \link[graphics]{graphical parameters}
    can also be specified as arguments.  See \sQuote{Details} for the
    interpretation of the default for \code{log}.

    For the \code{"function"} method of \code{plot}, \code{\dots} can
    include any of the other arguments of \code{curve}, except
    \code{expr}.
  }

\item{log}{labels and \link[graphics]{graphical parameters}
    can also be specified as arguments.  See \sQuote{Details} for the
    interpretation of the default for \code{log}.

    For the \code{"function"} method of \code{plot}, \code{\dots} can
    include any of the other arguments of \code{curve}, except
    \code{expr}.
  }

\item{xlim}{\code{NULL} or a numeric vector of length 2;
    if non-\code{NULL} it provides the defaults for \code{c(from, to)}
    and, unless \code{add = TRUE}, selects the x-limits of the plot -- see
    \code{\link[graphics]{plot.window}}.}

\item{grid}{logical; if \code{TRUE}, a background grid will be drawn}

\item{col.grid}{\code{grid} color}

\item{col.acc}{\code{grid} accent color}

\item{x}{a \sQuote{vectorizing} numeric \R function.}

\item{y}{alias for \code{from} for compatibility with \code{plot}}
}
\description{
Draws a curve corresponding to a function over the interval \code{[from, to]}.
\code{gcurve} can plot also an expression in the variable \code{xname}, 
default \code{x}.
}
