% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{get_palette}
\alias{get_palette}
\title{get_palette}
\usage{
get_palette(palette, n = 9, direction = 1)
}
\arguments{
\item{palette}{A vector of hexcode colors, or palettes from "RColorBrewer" or "viridis".}

\item{n}{Number of colors to return.}

\item{direction}{Direction of the palette vector (1 for forward, -1 for reverse).}
}
\value{
A vector of hexadecimal color codes.
}
\description{
Get a vector of hexa color codes.
}
