% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-predictAdditionalData.R
\name{getCommonPredictions}
\alias{getCommonPredictions}
\title{Predictions for Selected Models for a Common X Range}
\usage{
getCommonPredictions(activeDataList, xRange, isTest = FALSE)
}
\arguments{
\item{activeDataList}{(reactive) list of selected saved data}

\item{xRange}{(numeric) range for which correlation will be calculated}

\item{isTest}{TRUE if used within tests}
}
\description{
Predictions for Selected Models for a Common X Range
}
