% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-predictAdditionalData.R
\name{getCommonX}
\alias{getCommonX}
\title{Predictions for Selected Models for a Common X Range}
\usage{
getCommonX(activeDataList, what)
}
\arguments{
\item{activeDataList}{(reactive) list of selected saved data}

\item{what}{(character) name of the function to apply to xRange}
}
\description{
Predictions for Selected Models for a Common X Range
}
