% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-fitModel.R
\name{getLLog}
\alias{getLLog}
\title{Get LLog}
\usage{
getLLog(matrixDiff, XX, betamcSigma, smcSmpls, sdVar)
}
\arguments{
\item{matrixDiff}{(matrix) matrix difference}

\item{XX}{(matrix) design matrix}

\item{betamcSigma}{(matrix) coefficients of varying standard deviation spline}

\item{smcSmpls}{(numeric)  sigmas of MC samples}

\item{sdVar}{(logical) TRUE if variable standard deviation}
}
\description{
Log likelihood computation for (W)AIC
}
