% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-fitModel.R
\name{getPlotValues}
\alias{getPlotValues}
\title{Get plotValues}
\usage{
getPlotValues(
  plotValues,
  activeFile,
  activeFileData,
  dataSelection,
  modelParameters
)
}
\arguments{
\item{plotValues}{(list) list with all values required for the plot}

\item{activeFile}{(character) name of the selected file}

\item{activeFileData}{(data.frame) content of the selected file}

\item{dataSelection}{(list) list of reactive selected columns of the
current file}

\item{modelParameters}{(list) list of settings for the model}
}
\description{
Get plotValues
}
