% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-plotFunctions.R
\name{getRelevantOverlap}
\alias{getRelevantOverlap}
\title{Get Relevant Overlap}
\usage{
getRelevantOverlap(valuesOfQ, referenceCurve, sigLevel)
}
\arguments{
\item{valuesOfQ}{predictions of the curve q}

\item{referenceCurve}{predictions of the reference curve}

\item{sigLevel}{significance level}
}
\description{
Get Relevant Overlap
}
