% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplots.R
\name{ggmultiplot}
\alias{ggmultiplot}
\title{Draw multiple ggplot objects in a single layout}
\usage{
ggmultiplot(..., plotlist = NULL, ncol = 1L, layout = NULL)
}
\arguments{
\item{...}{ggplot objects}

\item{plotlist}{list of ggplot objects (optional)}

\item{ncol}{number of columns in layout}

\item{layout}{matrix specifying the layout; if present, \code{cols} is 
ignored; 

if layout is \code{matrix(c(1, 2, 3, 3), nrow = 2, byrow = TRUE)},
for example, then plot 1 will go in the upper left, 2 will go in the upper 
right, and 3 will go all the way across the bottom}
}
\description{
Uses functions in \pkg{grid} to arrange one or more 
\code{\link[ggplot2]{ggplot}} objects into a single layout.
}
\examples{
library('ggplot2')
gg <- ggplot(mtcars, aes(factor(1), fill = factor(cyl))) +
  geom_bar(width = 1) + theme_minimal()
gg1 <- gg + coord_polar()
gg2 <- gg + coord_polar(theta = 'y')
gg3 <- ggplot(mtcars, aes(factor(cyl), fill = factor(cyl))) +
  geom_bar(width = 1) + coord_polar() + theme_minimal()

ggmultiplot(gg)
ggmultiplot(gg1, gg2, gg3, ncol = 2)
ggmultiplot(
  gg1, gg2, gg3,
  layout = matrix(c(1, 2, 3, 3, 3, 3), ncol = 2, byrow = TRUE)
)
  
}
