% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_functions.R
\name{ggplot_col_from_vector}
\alias{ggplot_col_from_vector}
\title{Make a bar chart from a named numeric vector or table}
\usage{
ggplot_col_from_vector(
  vec,
  vec_levels = NULL,
  xlab = NULL,
  ylab = NULL,
  title = NULL,
  x_labels_angle = NULL
)
}
\arguments{
\item{vec}{A named numeric vector}

\item{vec_levels}{The order of the names in the plot}

\item{xlab}{The x-axis title}

\item{ylab}{The y-axis title}

\item{title}{The title of the plot}

\item{x_labels_angle}{The angle to rotate the x-axis labels.}
}
\value{
A ggplot object
}
\description{
Make a bar chart from a named numeric vector or table
}
