% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_functions.R
\name{ggsave_list}
\alias{ggsave_list}
\title{Save a list of plots to separate files in a specified directory}
\usage{
ggsave_list(
  plotlist,
  filenames = NULL,
  file_format = "jpeg",
  directory = ".",
  width = 16,
  height = 9
)
}
\arguments{
\item{plotlist}{A list of ggplot objects}

\item{filenames}{A vector of filenames to use to save plots. Default is to use the variable names used for the plots.}

\item{file_format}{Extension of file format to use if filenames not provided. Default is "jpeg".}

\item{directory}{Directory to save images. Default is current directory.}

\item{width}{Width of the plots in inches}

\item{height}{Height of the plots in inches}
}
\description{
Save a list of plots to separate files in a specified directory
}
