% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplots.R
\name{ghist.Date}
\alias{ghist.Date}
\alias{ghist.POSIXt}
\title{ghistogram of a date or date-time object}
\usage{
\method{ghist}{Date}(
  x,
  breaks,
  ...,
  xlab = deparse(substitute(x)),
  plot = TRUE,
  freq = FALSE,
  start.on.monday = TRUE,
  format,
  grid = TRUE,
  col.grid = "grey90",
  col.acc = "white"
)

\method{ghist}{POSIXt}(
  x,
  breaks,
  ...,
  xlab = deparse(substitute(x)),
  plot = TRUE,
  freq = FALSE,
  start.on.monday = TRUE,
  format,
  grid = TRUE,
  col.grid = "grey90",
  col.acc = "white"
)
}
\arguments{
\item{x}{an object inheriting from class \code{"POSIXt"} or \code{"Date"}.}

\item{breaks}{
    a vector of cut points \emph{or} number giving the number of
    intervals which \code{x} is to be cut into \emph{or} an
    interval specification, one of \code{"days"}, \code{"weeks"},
    \code{"months"}, \code{"quarters"} or \code{"years"},
    plus \code{"secs"}, \code{"mins"}, \code{"hours"} for
    date-time objects.
 }

\item{...}{\link[graphics]{graphical parameters}, or arguments to
    \code{\link[graphics]{hist.default}} such as \code{include.lowest},
      \code{right} and \code{labels}.}

\item{xlab}{a character string giving the label for the x axis, if plotted.}

\item{plot}{logical.  If \code{TRUE} (default), a histogram is
    plotted, otherwise a list of breaks and counts is returned.}

\item{freq}{logical; if \code{TRUE}, the
    histogram graphic is a representation of frequencies, i.e,
    the \code{counts} component of the result; if \code{FALSE},
    \emph{relative} frequencies (probabilities) are plotted.}

\item{start.on.monday}{logical. If \code{breaks = "weeks"}, should the
    week start on Mondays or Sundays?}

\item{format}{for the x-axis labels. See \code{\link{strptime}}.}

\item{grid}{logical; if \code{TRUE}, a background grid will be drawn}

\item{col.grid}{\code{grid} color}

\item{col.acc}{\code{grid} accent color}
}
\description{
Method for \code{\link{ghist}} applied to date or date-time objects.
}
