% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplots.R
\name{gmatplot}
\alias{gmatplot}
\alias{gmatpoints}
\alias{gmatlines}
\title{Plos columns of matrices}
\usage{
gmatplot(x, y, type = "p", lty = 1:5, lwd = 1, lend = par("lend"),
  pch = NULL, col = 1:6, cex = NULL, bg = NA, xlab = NULL,
  ylab = NULL, xlim = NULL, ylim = NULL, ..., add = FALSE,
  verbose = getOption("verbose"), grid = TRUE, col.grid = "grey90",
  col.acc = "white")

gmatpoints(x, y, type = "p", lty = 1:5, lwd = 1, pch = NULL,
  col = 1:6, ..., grid = TRUE, col.grid = "grey90", col.acc = "white")

gmatlines(x, y, type = "l", lty = 1:5, lwd = 1, pch = NULL, col = 1:6,
  ..., grid = TRUE, col.grid = "grey90", col.acc = "white")
}
\arguments{
\item{x}{vectors or matrices of data for plotting.  The number of
    rows should match.  If one of them are missing, the other is taken
    as \code{y} and an \code{x} vector of \code{1:n} is used.
    Missing values (\code{NA}s) are allowed.}

\item{y}{vectors or matrices of data for plotting.  The number of
    rows should match.  If one of them are missing, the other is taken
    as \code{y} and an \code{x} vector of \code{1:n} is used.
    Missing values (\code{NA}s) are allowed.}

\item{type}{character string (length 1 vector) or vector of 1-character
    strings indicating the type of plot for each
    column of \code{y}, see \code{\link{plot}} for all possible
    \code{type}s.  The first character of \code{type}
    defines the first plot, the second character the second, etc.
    Characters in \code{type} are cycled through; e.g., \code{"pl"}
    alternately plots points and lines.}

\item{lty}{vector of line types, widths, and end styles.
    The first element is for the first column, the second element for
    the second column, etc., even if lines are not plotted for all
    columns. Line types will be used cyclically until all plots are
    drawn.}

\item{lwd}{vector of line types, widths, and end styles.
    The first element is for the first column, the second element for
    the second column, etc., even if lines are not plotted for all
    columns. Line types will be used cyclically until all plots are
    drawn.}

\item{lend}{vector of line types, widths, and end styles.
    The first element is for the first column, the second element for
    the second column, etc., even if lines are not plotted for all
    columns. Line types will be used cyclically until all plots are
    drawn.}

\item{pch}{character string or vector of 1-characters or integers for
    plotting characters, see \code{\link{points}}.
    The first character is the plotting-character for the first plot,
    the second for the second, etc.  The default is the digits (1
    through 9, 0) then the lowercase and uppercase letters.
\%\%__FIXME__ Which of the following is still true; what's the "right" UI ?
\%  To mix special plotting symbols (those specified by integers) and
\%  plotting characters, either use character escapes \code{pch = "Xe002"}
\%  or figure out the numeric equivalent of the ASCII character,
\%  \code{pch = c(88, 2)}.
  }

\item{col}{vector of colors.  Colors are used cyclically.}

\item{cex}{vector of character expansion sizes, used cyclically.
    This works as a multiple of \code{\link{par}("cex")}.  \code{NULL} is
    equivalent to \code{1.0}. }

\item{bg}{vector of background (fill) colors for the open plot
    symbols given by \code{pch = 21:25} as in \code{\link{points}}.  The
    default \code{NA} corresponds to the one of the underlying function
    \code{\link{plot.xy}}.}

\item{xlab}{titles for x and y axes, as in \code{\link{plot}}.}

\item{ylab}{titles for x and y axes, as in \code{\link{plot}}.}

\item{xlim}{ranges of x and y axes, as in \code{\link{plot}}.}

\item{ylim}{ranges of x and y axes, as in \code{\link{plot}}.}

\item{...}{Graphical parameters (see \code{\link{par}}) and any further
    arguments of \code{plot}, typically \code{\link{plot.default}}, may also be
    supplied as arguments to this function.  Hence, the high-level
    graphics control arguments described under \code{\link{par}} and the
    arguments to \code{\link{title}} may be supplied to this function.}

\item{add}{logical.  If \code{TRUE}, plots are added to current one,
    using \code{\link{points}} and \code{\link{lines}}.}

\item{verbose}{logical.  If \code{TRUE}, write one line of what is
    done.}

\item{grid}{logical; if \code{TRUE}, a background grid will be drawn}

\item{col.grid}{\code{grid} color}

\item{col.acc}{\code{grid} accent color}
}
\description{
Plot the columns of one matrix against the columns of another.
}
