% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplots.R
\name{gpairs}
\alias{gpairs}
\alias{gpairs.default}
\alias{gpairs.formula}
\title{Scatterplot matrices}
\usage{
gpairs(x, ...)

\method{gpairs}{default}(
  x,
  labels,
  panel = points,
  ...,
  lower.panel = panel,
  upper.panel = panel,
  diag.panel = NULL,
  text.panel = textPanel,
  label.pos = 0.5 + has.diag/3,
  line.main = 3,
  cex.labels = NULL,
  font.labels = 1,
  row1attop = TRUE,
  gap = 1,
  log = "",
  grid = TRUE,
  col.grid = "grey90",
  col.acc = "white"
)

\method{gpairs}{formula}(
  formula,
  data = NULL,
  ...,
  subset,
  na.action = stats::na.pass,
  grid = TRUE,
  col.grid = "grey90",
  col.acc = "white"
)
}
\arguments{
\item{x}{the coordinates of points given as numeric columns of a
    matrix or data frame.  Logical and factor columns are converted to
    numeric in the same way that \code{\link{data.matrix}} does.
  }

\item{...}{
    arguments to be passed to or from methods.

    Also, \link[graphics]{graphical parameters} can be given as can arguments to
    \code{plot} such as \code{main}.  \code{par("oma")} will be set
    appropriately unless specified.
  }

\item{labels}{the names of the variables.}

\item{panel}{\code{function(x, y, \dots)} which is used to plot
    the contents of each panel of the display.}

\item{lower.panel}{separate panel functions (or
    \code{NULL}) to be used below and above the diagonal respectively.}

\item{upper.panel}{separate panel functions (or
    \code{NULL}) to be used below and above the diagonal respectively.}

\item{diag.panel}{optional \code{function(x, \dots)} to be
    applied on the diagonals.}

\item{text.panel}{optional \code{function(x, y, labels, cex,
      font, \dots)} to be applied on the diagonals.}

\item{label.pos}{\code{y} position of labels in the text panel.}

\item{line.main}{if \code{main} is specified, \code{line.main} gives
    the \code{line} argument to \code{\link[graphics]{mtext}()} which draws the
    title.  You may want to specify \code{oma} when changing \code{line.main}.}

\item{cex.labels}{graphics parameters for the text panel.}

\item{font.labels}{graphics parameters for the text panel.}

\item{row1attop}{logical.  Should the layout be matrix-like with row 1
    at the top, or graph-like with row 1 at the bottom?  The latter (non
    default) leads to a basically symmetric scatterplot matrix.}

\item{gap}{distance between subplots, in margin lines.}

\item{log}{a character string indicating if logarithmic axes are to be
    used, see \code{\link[graphics]{plot.default}} \emph{or} a numeric vector of
    indices specifying the indices of those variables where logarithmic
    axes should be used for both x and y.  \code{log = "xy"} specifies
    logarithmic axes for all variables.}

\item{grid}{logical; if \code{TRUE}, a background grid will be drawn}

\item{col.grid}{\code{grid} color}

\item{col.acc}{\code{grid} accent color}

\item{formula}{a formula, such as \code{~ x + y + z}.  Each term will
    give a separate variable in the pairs plot, so terms should be
    numeric vectors.  (A response will be interpreted as another
    variable, but not treated specially, so it is confusing to use one.)}

\item{data}{a data.frame (or list) from which the variables in
    \code{formula} should be taken.}

\item{subset}{an optional vector specifying a subset of observations
    to be used for plotting.}

\item{na.action}{a function which indicates what should happen
    when the data contain \code{NA}s.  The default is to pass missing
    values on to the panel functions, but \code{na.action = na.omit}
    will cause cases with missing values in any of the variables to be
    omitted entirely.}
}
\description{
A matrix of scatterplots is produced.
}
