% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplots.R
\name{gplot.density}
\alias{gplot.density}
\title{gplot method for kernel density estimation}
\usage{
\method{gplot}{density}(x, main = NULL, xlab = NULL, ylab = "Density",
  type = "l", zero.line = TRUE, grid = TRUE, col.grid = "grey90",
  col.acc = "white", ...)
}
\arguments{
\item{x}{a \code{"density"} object.}

\item{main}{plotting parameters with useful defaults.}

\item{xlab}{plotting parameters with useful defaults.}

\item{ylab}{plotting parameters with useful defaults.}

\item{type}{plotting parameters with useful defaults.}

\item{zero.line}{logical; if \code{TRUE}, add a base line at \eqn{y = 0}}

\item{grid}{logical; if \code{TRUE}, a background grid will be drawn}

\item{col.grid}{\code{grid} color}

\item{col.acc}{\code{grid} accent color}

\item{...}{further plotting parameters.}
}
\description{
The \code{gplot} method for density objects
}
