% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplots.R
\name{gplot.factor}
\alias{gplot.factor}
\title{Plotting factor variables}
\usage{
\method{gplot}{factor}(x, y, legend.text = NULL, ..., grid = TRUE,
  col.grid = "grey90", col.acc = "white")
}
\arguments{
\item{x}{numeric or factor. \code{y} may be missing.}

\item{legend.text}{character vector for annotation of y axis in the
      case of a factor \code{y}: defaults to \code{levels(y)}.  This
      sets the \code{yaxlabels} argument of \code{\link{spineplot}}.}

\item{...}{Further arguments to \code{\link{barplot}},
      \code{\link{boxplot}}, \code{\link{spineplot}} or
      \code{\link{plot}} as appropriate.  All of these accept
      \link{graphical parameters} (see \code{\link{par}}) and
      annotation arguments passed
      to \code{\link{title}} and \code{axes = FALSE}.  None accept \code{type}.
    }

\item{grid}{logical; if \code{TRUE}, a background grid will be drawn}

\item{col.grid}{\code{grid} color}

\item{col.acc}{\code{grid} accent color}
}
\description{
This function implements a scatterplot for \code{\link{factor}} arguments
for the \emph{generic} \code{\link{gplot}} function.

If \code{y} is missing \code{\link{gbarplot}} is produced; for numeric 
\code{y} \code{\link{gboxplot}} is used; for a factor \code{y} a 
\code{\link{gspineplot}} is shown; and for any other type of \code{y}, the
next \code{gplot} method is called, normally \code{\link{gplot.default}}.
}

