% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplots.R
\name{gplot.formula}
\alias{gplot.formula}
\title{Formula notation for gplot scatterplots}
\usage{
\method{gplot}{formula}(
  formula,
  data = parent.frame(),
  ...,
  subset,
  ylab = varnames[response],
  ask = dev.interactive(),
  grid = TRUE,
  col.grid = "grey90",
  col.acc = "white"
)
}
\arguments{
\item{formula}{a \code{\link{formula}}, such as \code{y ~ x}.}

\item{data}{a data.frame (or list) from which the variables in
    \code{formula} should be taken.  A matrix is converted to a data frame.}

\item{...}{Arguments to be passed to or from other methods.
    \code{horizontal = TRUE} is also accepted.}

\item{subset}{an optional vector specifying a subset of observations
    to be used in the fitting process.}

\item{ylab}{the y label of the plot(s).}

\item{ask}{logical, see \code{\link[graphics]{par}}.}

\item{grid}{logical; if \code{TRUE}, a background grid will be drawn}

\item{col.grid}{\code{grid} color}

\item{col.acc}{\code{grid} accent color}
}
\description{
Specify a \code{\link{gplot}} via a formula
}
