% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplots.R
\name{gplot.histogram}
\alias{gplot.histogram}
\title{gplot histograms}
\usage{
\method{gplot}{histogram}(x, freq = equidist, density = NULL, angle = 45,
  col = "white", border = par("fg"), lty = NULL,
  main = paste("Histogram of", paste(x$xname, collapse = "\\n")),
  sub = NULL, xlab = x$xname, ylab, xlim = range(x$breaks), ylim = NULL,
  axes = TRUE, labels = FALSE, add = FALSE, ann = TRUE, grid = TRUE,
  col.grid = "grey90", col.acc = "white", ...)
}
\arguments{
\item{x}{a \code{histogram} object, or a list with components
    \code{density}, \code{mid}, etc, see \code{\link{hist}} for
    information about the components of \code{x}.}

\item{freq}{logical; if \code{TRUE}, the
    histogram graphic is to present a representation of frequencies, i.e,
    \code{x$counts}; if \code{FALSE}, \emph{relative} frequencies
    (probabilities), i.e., \code{x$density}, are plotted.  The
    default is true for equidistant \code{breaks} and false otherwise.}

\item{density}{select shading of bars by lines:
    see \code{\link{rect}}.}

\item{angle}{select shading of bars by lines:
    see \code{\link{rect}}.}

\item{col}{a colour to be used to fill the bars.  The default of
    \code{NULL} yields unfilled bars.}

\item{border}{the color of the border around the bars.}

\item{lty}{the line type used for the bars, see also \code{\link{lines}}.}

\item{main}{these arguments to \code{title} have useful
    defaults here.}

\item{sub}{these arguments to \code{title} have useful
    defaults here.}

\item{xlab}{these arguments to \code{title} have useful
    defaults here.}

\item{ylab}{these arguments to \code{title} have useful
    defaults here.}

\item{xlim}{the range of x and y values with sensible defaults.}

\item{ylim}{the range of x and y values with sensible defaults.}

\item{axes}{logical, indicating if axes should be drawn.}

\item{labels}{logical or character.  Additionally draw labels on top of bars,
    if not \code{FALSE}; if \code{TRUE}, draw the counts or rounded
    densities; if \code{labels} is a \code{character}, draw itself.}

\item{add}{logical.  If \code{TRUE}, only the bars are added to the
    current plot.  This is what \code{lines.histogram(*)} does.}

\item{ann}{logical.  Should annotations (titles and axis titles) be plotted?}

\item{grid}{logical; if \code{TRUE}, a background grid will be drawn}

\item{col.grid}{\code{grid} color}

\item{col.acc}{\code{grid} accent color}

\item{...}{further \link{graphical parameters} to
    \code{title} and \code{axis}.}
}
\description{
These are methods for objects of class "\code{histogram}", typically 
produced by \code{\link{ghist}}.
}
