% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplots.R
\name{gplot.lm}
\alias{gplot.lm}
\title{gplot diagnostics for an lm object}
\usage{
\method{gplot}{lm}(x, which = c(1L:3L, 5L), grid = TRUE,
  col.grid = "grey90", col.acc = "white",
  caption = list("Residuals vs Fitted", "Normal Q-Q", "Scale-Location",
  "Cook's distance", "Residuals vs Leverage",
  expression("Cook's dist vs Leverage  " * h[ii]/(1 - h[ii]))), panel = if
  (add.smooth) panel.smooth else points, sub.caption = NULL, main = "",
  ask = prod(par("mfcol")) < length(which) && dev.interactive(), ...,
  id.n = 3, labels.id = names(residuals(x)), cex.id = 0.75,
  qqline = TRUE, cook.levels = c(0.5, 1),
  add.smooth = getOption("add.smooth"), label.pos = c(4, 2),
  cex.caption = 1)
}
\arguments{
\item{x}{\code{lm} object, typically result of \code{\link{lm}} or
    \code{\link{glm}}.}

\item{which}{if a subset of the plots is required, specify a subset of
    the numbers \code{1:6}.}

\item{grid}{logical; if \code{TRUE}, a background grid will be drawn}

\item{col.grid}{\code{grid} color}

\item{col.acc}{\code{grid} accent color}

\item{caption}{captions to appear above the plots;
    \code{\link{character}} vector or \code{\link{list}} of valid
    graphics annotations, see \code{\link{as.graphicsAnnot}}, of length
    6, the j-th entry corresponding to \code{which[j]}.  Can be set to
    \code{""} or \code{NA} to suppress all captions.
  }

\item{panel}{panel function.  The useful alternative to
    \code{\link{points}}, \code{\link{panel.smooth}} can be chosen
    by \code{add.smooth = TRUE}.}

\item{sub.caption}{common title---above the figures if there are more
    than one; used as \code{sub} (s.\code{\link{title}}) otherwise.  If
    \code{NULL}, as by default, a possible abbreviated version of
    \code{deparse(x$call)} is used.}

\item{main}{title to each plot---in addition to \code{caption}.}

\item{ask}{logical; if \code{TRUE}, the user is \emph{ask}ed before
    each plot, see \code{\link{par}(ask=.)}.}

\item{...}{other parameters to be passed through to plotting
    functions.}

\item{id.n}{number of points to be labelled in each plot, starting
    with the most extreme.}

\item{labels.id}{vector of labels, from which the labels for extreme
    points will be chosen.  \code{NULL} uses observation numbers.}

\item{cex.id}{magnification of point labels.}

\item{qqline}{logical indicating if a \code{\link{qqline}()} should be
    added to the normal Q-Q plot.}

\item{cook.levels}{levels of Cook's distance at which to draw contours.}

\item{add.smooth}{logical indicating if a smoother should be added to
    most plots; see also \code{panel} above.}

\item{label.pos}{positioning of labels, for the left half and right
    half of the graph respectively, for plots 1-3.}

\item{cex.caption}{controls the size of \code{caption}.}
}
\description{
Six plots (selectable by \code{which}) are currently available: a plot of
residuals against fitted values, a Scale-Location plot of 
\emph{sqrt(| residuals |)} against fitted values, a Normal Q-Q plot, a 
plot of Cook's distances versus row labels, a plot of residuals against 
leverages, and a plot of Cook's distances against leverage/(1-leverage). 
By default, the first three and 5 are provided.
}

