% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gplot.table}
\alias{gplot.table}
\title{gplot method for \code{table} objects}
\usage{
\method{gplot}{table}(x, type = "h", ylim = c(0, max(x)), lwd = 2,
  xlab = NULL, ylab = NULL, frame.plot = is.num, ..., grid = TRUE,
  col.grid = "grey90", col.acc = "white")
}
\arguments{
\item{x}{a \code{\link{table}} (like) object.}

\item{type}{plotting type.}

\item{ylim}{range of y-axis.}

\item{lwd}{line width for bars when \code{type = "h"} is used in the 1D case.}

\item{xlab}{x- and y-axis labels.}

\item{frame.plot}{logical indicating if a frame (\code{\link{box}})
    should be drawn in the 1D case.  Defaults to true when \code{x} has
    \code{\link{dimnames}} coerce-able to numbers.}

\item{...}{further graphical arguments, see \code{\link{plot.default}}.
    \code{axes = FALSE} is accepted.}

\item{grid}{logical; if \code{TRUE}, a background grid will be drawn}

\item{col.grid}{\code{grid} color}

\item{col.acc}{\code{grid} accent color}
}
\description{
This is a method of the generic \code{gplot} function for (contingency)
\code{\link{table}} objects. Whereas for two- and more dimensional tables,
a \code{\link{mosaicplot}} is drawn, one-dimensional ones are plotted as
bars.
}

