% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplots.R
\name{gqqnorm}
\alias{gqqnorm}
\alias{gqqnorm.default}
\alias{gqqplot}
\title{Quantile-quantile gplots}
\usage{
gqqnorm(y, ...)

\method{gqqnorm}{default}(y, ylim, main = "Normal Q-Q Plot",
  xlab = "Theoretical Quantiles", ylab = "Sample Quantiles",
  plot.it = TRUE, datax = FALSE, grid = TRUE, col.grid = "grey90",
  col.acc = "white", ...)

gqqplot(x, y, plot.it = TRUE, xlab = deparse(substitute(x)),
  ylab = deparse(substitute(y)), ..., grid = TRUE, col.grid = "grey90",
  col.acc = "white")
}
\arguments{
\item{y}{The second or only data sample.}

\item{ylim}{graphical parameters.}

\item{main}{plot labels.  The \code{xlab} and \code{ylab}
    refer to the y and x axes respectively if \code{datax = TRUE}.}

\item{xlab}{plot labels.  The \code{xlab} and \code{ylab}
    refer to the y and x axes respectively if \code{datax = TRUE}.}

\item{ylab}{plot labels.  The \code{xlab} and \code{ylab}
    refer to the y and x axes respectively if \code{datax = TRUE}.}

\item{plot.it}{logical. Should the result be plotted?}

\item{datax}{logical. Should data values be on the x-axis?}

\item{grid}{logical; if \code{TRUE}, a background grid will be drawn}

\item{col.grid}{\code{grid} color}

\item{col.acc}{\code{grid} accent color}

\item{x}{The first sample for \code{qqplot}.}
}
\description{
qqnorm is a generic function the default method of which produces a normal
QQ plot of the values in \code{y}. \code{qqline} adds a line to a 
“theoretical”, by default normal, quantile-quantile plot which passes 
through the \code{probs} quantiles, by default the first and third quartiles.

\code{qqplot} produces a QQ plot of two datasets.

Graphical parameters may be given as arguments to \code{gqqnorm}, 
\code{gqqplot} and \code{qqline}.
}
