% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gspineplot}
\alias{gspineplot}
\alias{gspineplot.default}
\alias{gspineplot.formula}
\title{Spine plots and spinograms}
\usage{
gspineplot(x, ...)

\method{gspineplot}{default}(x, y = NULL, breaks = NULL, tol.ylab = 0.05,
  off = NULL, ylevels = NULL, col = NULL, main = "", xlab = NULL,
  ylab = NULL, xaxlabels = NULL, yaxlabels = NULL, xlim = NULL,
  ylim = c(0, 1), axes = TRUE, ..., grid = TRUE, col.grid = "grey90",
  col.acc = "white")

\method{gspineplot}{formula}(formula, data = NULL, breaks = NULL,
  tol.ylab = 0.05, off = NULL, ylevels = NULL, col = NULL, main = "",
  xlab = NULL, ylab = NULL, xaxlabels = NULL, yaxlabels = NULL,
  xlim = NULL, ylim = c(0, 1), axes = TRUE, ..., subset = NULL,
  grid = TRUE, col.grid = "grey90", col.acc = "white")
}
\arguments{
\item{x}{an object, the default method expects either a single variable
    (interpreted to be the explanatory variable) or a 2-way table.  See
    details.}

\item{...}{additional arguments passed to \code{\link{rect}}.}

\item{y}{a \code{"factor"} interpreted to be the dependent variable}

\item{breaks}{if the explanatory variable is numeric, this controls how
    it is discretized. \code{breaks} is passed to \code{\link{hist}} and can
    be a list of arguments.}

\item{tol.ylab}{convenience tolerance parameter for y-axis annotation.
    If the distance between two labels drops under this threshold, they are
    plotted equidistantly.}

\item{off}{vertical offset between the bars (in per cent). It is fixed to
    \code{0} for spinograms and defaults to \code{2} for spine plots.}

\item{ylevels}{a character or numeric vector specifying in which order
    the levels of the dependent variable should be plotted.}

\item{col}{a vector of fill colors of the same length as \code{levels(y)}.
    The default is to call \code{\link{gray.colors}}.}

\item{main}{character strings for annotation}

\item{xaxlabels}{character vectors for annotation of x and y axis.
    Default to \code{levels(y)} and \code{levels(x)}, respectively for the
    spine plot. For \code{xaxlabels} in the spinogram, the breaks are
    used.}

\item{xlim}{the range of x and y values with sensible defaults.}

\item{axes}{logical.  If \code{FALSE} all axes (including those giving
    level names) are suppressed.}

\item{grid}{logical; if \code{TRUE}, a background grid will be drawn}

\item{col.grid}{\code{grid} color}

\item{col.acc}{\code{grid} accent color}

\item{formula}{a \code{"formula"} of type \code{y ~ x} with a single
    dependent \code{"factor"} and a single explanatory variable.}

\item{data}{an optional data frame.}

\item{subset}{an optional vector specifying a subset of observations
    to be used for plotting.  }
}
\description{
Spine plots are a special case of mosaic plots and can be seen as a
generalization of stacked (or highlighted) bar plots. Analogously, spinograms
are an extension of histograms.
}

