% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gplots.R
\name{gstripchart}
\alias{gstripchart}
\alias{gstripchart.default}
\alias{gstripchart.formula}
\title{1-D scatter plots}
\usage{
gstripchart(x, ...)

\method{gstripchart}{default}(x, method = "overplot", jitter = 0.1,
  offset = 1/3, vertical = FALSE, group.names, add = FALSE, at = NULL,
  xlim = NULL, ylim = NULL, ylab = NULL, xlab = NULL, dlab = "",
  glab = "", log = "", pch = 0, col = par("fg"), cex = par("cex"),
  axes = TRUE, frame.plot = axes, ..., grid = TRUE, col.grid = "grey90",
  col.acc = "white")

\method{gstripchart}{formula}(x, data = NULL, dlab = NULL, ..., subset,
  na.action = NULL, grid = TRUE, col.grid = "grey90", col.acc = "white")
}
\arguments{
\item{x}{the data from which the plots are to be produced.  In the
    default method the data can be specified as a single numeric
    vector, or as list of numeric vectors, each corresponding to
    a component plot.  In the \code{formula} method, a symbolic
    specification of the form \code{y ~ g} can be given,
    indicating the observations in the vector \code{y} are to be
    grouped according to the levels of the factor
    \code{g}.  \code{NA}s are allowed in the data.}

\item{...}{additional parameters passed to the default method, or by
    it to \code{\link{plot.window}}, \code{\link{points}},
    \code{\link{axis}} and \code{\link{title}} to control the appearance
    of the plot.}

\item{method}{the method to be used to separate coincident points.
    The default method \code{"overplot"} causes such points to be
    overplotted, but it is also possible to specify \code{"jitter"} to
    jitter the points, or \code{"stack"} have coincident points
    stacked.  The last method only makes sense for very granular data.}

\item{jitter}{when \code{method = "jitter"} is used, \code{jitter}
    gives the amount of jittering applied.}

\item{offset}{when stacking is used, points are stacked this many
    line-heights (symbol widths) apart.}

\item{vertical}{when vertical is \code{TRUE} the plots are drawn
    vertically rather than the default horizontal.}

\item{group.names}{group labels which will be printed alongside
    (or underneath) each plot.}

\item{add}{logical, if true \emph{add} the chart to the current plot.}

\item{at}{numeric vector giving the locations where the charts should
    be drawn, particularly when \code{add = TRUE};
    defaults to \code{1:n} where \code{n} is the number of boxes.}

\item{xlim}{plot limits: see \code{\link{plot.window}}.}

\item{ylab}{labels: see \code{\link{title}}.}

\item{dlab}{alternate way to specify axis labels: see \sQuote{Details}.}

\item{log}{on which axes to use a log scale:  see
    \code{\link{plot.default}}}

\item{pch}{Graphical parameters: see \code{\link{par}}.}

\item{axes}{Axis control:  see \code{\link{plot.default}}.}

\item{grid}{logical; if \code{TRUE}, a background grid will be drawn}

\item{col.grid}{\code{grid} color}

\item{col.acc}{\code{grid} accent color}

\item{data}{a data.frame (or list) from which the variables in
    \code{x} should be taken.}

\item{subset}{an optional vector specifying a subset of observations
    to be used for plotting.}

\item{na.action}{a function which indicates what should happen
    when the data contain \code{NA}s.  The default is to ignore missing
    values in either the response or the group.}
}
\description{
\code{gstripchart} produces one-dimensional scatter plots (or dot plots) of
the given data. These plots are a good alternative to \code{\link{gboxplot}s}
when the sample sizes are small.
}

