% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gplots.R
\name{gsunflowerplot}
\alias{gsunflowerplot}
\alias{gsunflowerplot.default}
\alias{gsunflowerplot.formula}
\title{Produce a gsunflower scatter plot}
\usage{
gsunflowerplot(x, ...)

\method{gsunflowerplot}{default}(x, y = NULL, number, log = "",
  digits = 6L, xlab = NULL, ylab = NULL, xlim = NULL, ylim = NULL,
  add = FALSE, rotate = FALSE, pch = 16, cex = 0.8, cex.fact = 1.5,
  col = par("col"), bg = NA, size = 1/8, seg.col = 2, seg.lwd = 1.5,
  ..., grid = TRUE, col.grid = "grey90", col.acc = "white")

\method{gsunflowerplot}{formula}(formula, data = NULL, xlab = NULL,
  ylab = NULL, ..., subset, na.action = NULL, grid = TRUE,
  col.grid = "grey90", col.acc = "white")
}
\arguments{
\item{x}{numeric vector of \code{x}-coordinates of length \code{n},
    say, or another valid plotting structure, as for
    \code{\link{plot.default}}, see also \code{\link{xy.coords}}.}

\item{...}{further arguments to \code{\link{plot}} [if
    \code{add = FALSE}], or to be passed to or from another method.}

\item{y}{numeric vector of \code{y}-coordinates of length \code{n}.}

\item{number}{integer vector of length \code{n}. \code{number[i]} = number
    of replicates for \code{(x[i], y[i])}, may be 0.\cr
    Default (\code{missing(number)}): compute the exact multiplicity of
    the points \code{x[], y[]}, via
    \code{\link{xyTable}()}.}

\item{log}{character indicating log coordinate scale, see
    \code{\link{plot.default}}.}

\item{digits}{when \code{number} is computed (i.e., not specified),
    \code{x} and \code{y} are rounded to \code{digits} significant
    digits before multiplicities are computed.}

\item{xlab}{character label for x-, or y-axis, respectively.}

\item{xlim}{\code{numeric(2)} limiting the extents of the x-,
    or y-axis.}

\item{add}{logical; should the plot be added on a previous one ?
    Default is \code{FALSE}.}

\item{rotate}{logical; if \code{TRUE}, randomly rotate the
    sunflowers (preventing artefacts).}

\item{pch}{plotting character to be used for points
    (\code{number[i]==1}) and center of sunflowers.}

\item{cex}{numeric; character size expansion of center points
    (s. \code{pch}).}

\item{cex.fact}{numeric \emph{shrinking} factor to be used for the
    center points \emph{when there are flower leaves},
    i.e., \code{cex / cex.fact} is used for these.}

\item{col}{colors for the plot symbols, passed to
    \code{\link{plot.default}}.}

\item{size}{of sunflower leaves in inches, 1[in] := 2.54[cm].
    Default: 1/8\", approximately 3.2mm.}

\item{seg.col}{color to be used for the \bold{seg}ments which make the
    sunflowers leaves, see \code{\link{par}(col=)};
    \code{col = "gold"} reminds of real sunflowers.}

\item{seg.lwd}{numeric; the line width for the leaves' segments.}

\item{grid}{logical; if \code{TRUE}, a background grid will be drawn}

\item{col.grid}{\code{grid} color}

\item{col.acc}{\code{grid} accent color}

\item{formula}{a \code{\link{formula}}, such as \code{y ~ x}.}

\item{data}{a data.frame (or list) from which the variables in
    \code{formula} should be taken.}

\item{subset}{an optional vector specifying a subset of observations
    to be used in the fitting process.}

\item{na.action}{a function which indicates what should happen
    when the data contain \code{NA}s.  The default is to ignore case
    with missing values.}
}
\description{
Multiple points are plotted as "sunflowers" with multiple leaves ("petals")
such that overplotting is visualised instead of accidental and invisible.
}

