% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{heatmap_without_clustering}
\alias{heatmap_without_clustering}
\title{Create a heatmap from a matrix without clustering}
\usage{
heatmap_without_clustering(
  mat,
  colors = colorRampPalette(RColorBrewer::brewer.pal(n = 7, name = "Reds"))(100),
  reverse_colors = F,
  breaks = NA,
  mask_diagonal = T,
  decimal_places = 2,
  title = "",
  row_labels = NULL,
  col_labels = NULL,
  angle_col = 270,
  title_size = 18,
  label_size = 15,
  number_size = 18,
  filename = NA,
  return_ggplot = F
)
}
\arguments{
\item{mat}{A matrix or object which can ne coerced to a matrix.}

\item{colors}{A vector if colors to use for heatmap. Default is colorRampPalette(RColorBrewer::brewer.pal(n = 7, name = "Reds"))(100).
Can also just supply a single colour and it will be interpolated using colorRampPalette().}

\item{reverse_colors}{A logical value indicating whether to reverse the supplied or created colors e.g. if all values are negative. Default is FALSE.}

\item{breaks}{Vector of breaks associated with colors. Default is NA.}

\item{mask_diagonal}{Logical value indicating whether main diagonal should be masked. Default is T.}

\item{decimal_places}{Number of decimal places to display. Default is 2.}

\item{title}{Title of the plot. Default is no title.}

\item{row_labels}{Labels to use for rows. Default is to use row names of mat.}

\item{col_labels}{Labels to use for columns. Default is to use column names of mat.}

\item{angle_col}{Angle to rotate column labels. Default is 270.}

\item{title_size}{Fontsize for the title. Default is 18.}

\item{label_size}{Fontsize for the labels. Default is 15.}

\item{number_size}{Fontsize for numbers. Default is 18.}

\item{filename}{Optional filename to save plot with.}

\item{return_ggplot}{A logical value indicating whether the heatmap should be returned as a ggplot object}
}
\value{
A pheatmap object or ggplot object dericed from it depending on
}
\description{
Create a heatmap from a matrix without clustering
}
