% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-plotFunctions.R
\name{hex_with_opacity}
\alias{hex_with_opacity}
\title{Convert hex to hex with opacity}
\usage{
hex_with_opacity(hex, opacity)
}
\arguments{
\item{hex}{color in hex format}

\item{opacity}{opacity value between 0 and 1}
}
\description{
Convert hex to hex with opacity
}
