% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{histr}
\alias{histr}
\title{histr}
\usage{
histr(x, ..., lines.pars, rug.pars, poly.pars)
}
\arguments{
\item{x}{a vector of values for which the histogram is desired}

\item{...}{additional parameters passed to \code{\link{hist}} or graphical
parameters passed to \code{\link{par}}}

\item{rug.pars}{optional list of additional parameters passed to
\code{\link{rug}}}

\item{poly.pars}{optional list of additional parameters passed to
\code{\link{polygon}}}

\item{line.pars}{optional list of additional parameters passed to
\code{\link{line}}}
}
\value{
A list of length two containing the return value of \code{\link{hist}}
and \code{\link{density}} for \code{x}.
}
\description{
A histogram with density curve and rug.
}
\examples{
set.seed(1)
histr(x <- rnorm(100), xlim = c(-3, 3), las = 1)
plot(density(x), xlim = c(-3, 3))

histr(x + 50, main = 'Age at diagnosis', xlab = 'Age',
      poly.pars = list(col = 'dodgerblue2', density = 30),
      lines.pars = list(lty = 'dashed', lwd = 3),
      rug.pars = list(side = 3, col = 'red'))

}

