% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2.R
\name{inbar}
\alias{inbar}
\title{Inset barplot}
\usage{
inbar(
  height,
  width = 1,
  r = NULL,
  alpha = if (group.col) r else 1,
  space = NULL,
  names.arg = NULL,
  legend.text = NULL,
  horiz = FALSE,
  density = NULL,
  angle = 45,
  col = NULL,
  group.col = TRUE,
  border = par("fg"),
  main = NULL,
  sub = NULL,
  xlab = NULL,
  ylab = NULL,
  xlim = NULL,
  ylim = NULL,
  xpd = TRUE,
  log = "",
  axes = TRUE,
  axisnames = TRUE,
  cex.axis = par("cex.axis"),
  cex.names = par("cex.axis"),
  plot = TRUE,
  axis.lty = 0,
  offset = 0,
  add = FALSE,
  ann = !add && par("ann"),
  args.legend = NULL,
  ...
)
}
\arguments{
\item{height}{a matrix of data describing the bars which make up the plot,
usually a main/total bar and one or more smaller or subsets of the total;
bars are grouped by columns; the first row will be the main bar and other
rows will be minor bars}

\item{width}{optional vector of bar widths. Re-cycled to length the
    number of bars drawn.  Specifying a single value will have no
    visible effect unless \code{xlim} is specified.}

\item{r}{distance (in user coordinates) each inset bar is from the edge
of the main bar}

\item{alpha}{numeric value in \code{[0,1]} for the alpha transparency;
either a vector equal to \code{nrow(height)} (recycled as needed) or a
matrix having the same dimensions of \code{height}}

\item{space}{the amount of space (as a fraction of the average bar
    width) left before each bar.  May be given as a single number or
    one number per bar.  If \code{height} is a matrix and
    \code{beside} is \code{TRUE}, \code{space} may be specified by
    two numbers, where the first is the space between bars in the
    same group, and the second the space between the groups.  If not
    given explicitly, it defaults to \code{c(0,1)} if \code{height}
    is a matrix and \code{beside} is \code{TRUE}, and to 0.2
    otherwise.}

\item{names.arg}{a vector of names to be plotted below each bar or
    group of bars.  If this argument is omitted, then the names are
    taken from the \code{names} attribute of \code{height} if this
    is a vector, or the column names if it is a matrix.}

\item{legend.text}{a vector of text used to construct a legend for
    the plot, or a logical indicating whether a legend should be
    included.  This is only useful when \code{height} is a matrix.
    In that case given legend labels should correspond to the rows of
    \code{height}; if \code{legend.text} is true, the row names of
    \code{height} will be used as labels if they are non-null.}

\item{horiz}{a logical value.  If \code{FALSE}, the bars are drawn
    vertically with the first bar to the left.  If \code{TRUE}, the
    bars are drawn horizontally with the first at the bottom.}

\item{density}{a vector giving the density of shading lines, in
    lines per inch, for the bars or bar components.
    The default value of \code{NULL} means that no shading lines
    are drawn. Non-positive values of \code{density} also inhibit the
    drawing of shading lines.}

\item{angle}{the slope of shading lines, given as an angle in
    degrees (counter-clockwise), for the bars or bar components.}

\item{col}{colors for each bar; if a vector, each \code{col} is applied to
a group of bars with \code{alpha} transparency added to each minor bar;
alternatively, a matrix of colors having the same dimensions of
\code{height} to set each bar color}

\item{group.col}{logical; if \code{TRUE} (default), columns of \code{height}
are treated as groups and mapped to one color; if \code{FALSE}, each row
is mapped to \code{col} and recycled for each column of \code{height}}

\item{border}{the color to be used for the border of the bars.
    Use \code{border = NA} to omit borders.  If there are shading
    lines, \code{border = TRUE} means use the same colour for
    the border as for the shading lines.}

\item{main, sub}{main title and subtitle for the plot.}

\item{xlab}{a label for the x axis.}

\item{ylab}{a label for the y axis.}

\item{xlim}{limits for the x axis.}

\item{ylim}{limits for the y axis.}

\item{xpd}{logical. Should bars be allowed to go outside region?}

\item{log}{string specifying if axis scales should be logarithmic; see
    \code{\link[graphics]{plot.default}}.}

\item{axes}{logical.  If \code{TRUE}, a vertical (or horizontal, if
    \code{horiz} is true) axis is drawn.}

\item{axisnames}{logical.  If \code{TRUE}, and if there are
    \code{names.arg} (see above), the
    other axis is drawn (with \code{lty = 0}) and labeled.}

\item{cex.axis}{expansion factor for numeric axis labels (see
    \code{\link[graphics]{par}('cex')}).}

\item{cex.names}{expansion factor for axis names (bar labels).}

\item{plot}{logical.  If \code{FALSE}, nothing is plotted.}

\item{axis.lty}{the graphics parameter \code{lty} (see
    \code{\link[graphics]{par}('lty')}) applied to the axis
    and tick marks of the categorical (default horizontal) axis.  Note
    that by default the axis is suppressed.}

\item{offset}{a vector indicating how much the bars should be shifted
    relative to the x axis.}

\item{add}{logical specifying if bars should be added to an already
    existing plot; defaults to \code{FALSE}.}

\item{ann}{logical specifying if the default annotation (\code{main},
    \code{sub}, \code{xlab}, \code{ylab}) should appear on the plot, see
    \code{\link[graphics]{title}}.}

\item{args.legend}{list of additional arguments to pass to
    \code{\link[graphics]{legend}()}; names of the list are used as argument
    names.  Only used if \code{legend.text} is supplied.}

\item{...}{arguments to be passed to/from other methods.  For the
    default method these can include further arguments (such as
    \code{axes}, \code{asp} and \code{main}) and
    \link[graphics]{graphical parameters} (see \code{\link[graphics]{par}}) which are passed to
    \code{\link[graphics]{plot.window}()}, \code{\link[graphics]{title}()} and
    \code{\link[graphics]{axis}}.}
}
\value{
A numeric vector with the midpoint of each main bar (i.e., identical to a
call to \code{barplot(..., beside = FALSE)}).

Additionally, a list attribute (\code{attr(., "coords")}) where each list
element corresponds to a group of bars. The matrices contain coordinates
of each rectangle plus midpoints.
}
\description{
Draw a \code{\link{barplot}} with inset bars.
}
\examples{
set.seed(1)
tbl <- sapply(1:3, function(x) sort(rpois(3, 10), decreasing = TRUE))
barplot(tbl, col = 1:3)
inbar(tbl, col = 1:3)
inbar(tbl, col = matrix(rainbow(length(tbl)), nrow(tbl)))

## compare group.col = TRUE (default)
inbar(tbl, col = 1:3)
## compare group.col = FALSE
inbar(tbl, col = 1:3, group.col = FALSE)

inbar(
  tbl, col = 1:3, r = c(0.5, 0.8), alpha = c(1, 0.5, 0.25),
  horiz = TRUE, names.arg = 1:3,
  legend.text = 1:3, xlim = c(0, 20), border = NA
)

## inbar returns rect coordinates for future use
co <- inbar(tbl)
co <- do.call('rbind', attr(co, 'coords'))
points(co[, 'midpoint'], co[, 'ytop'], xpd = NA)

}
\seealso{
\code{\link{tracebar}}
}
