% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-uploadFiles.R
\name{incIndexOfFile}
\alias{incIndexOfFile}
\title{Inc Index Of File}
\usage{
incIndexOfFile(fileName)
}
\arguments{
\item{fileName}{(character) filename}
}
\description{
If the file has no index, add a new index: "(1)". If an index already exists, increase it by one.
}
