% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.color.R
\name{is.hex}
\alias{is.hex}
\title{Test HEX color validity}
\usage{
is.hex(color, digit6 = TRUE)
}
\arguments{
\item{color}{vector of colors to test}

\item{digit6}{if `TRUE`, tests for 6-digit HEX colors; otherwise, tests for 3-digit. `TRUE` by default.}
}
\description{
Check if all values in vector are a valid HEX
}
\keyword{Utilities}
