% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.envir.tools.R
\name{load.palette}
\alias{load.palette}
\title{Select color palette}
\usage{
load.palette(palette)
}
\arguments{
\item{palette}{name of color palette. Use `view.palette()` to see available color palettes.}
}
\description{
Loads a palette preset to use.
}
