% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{loess_smooth}
\alias{loess_smooth}
\alias{loess_smooth.formula}
\alias{loess_smooth.default}
\alias{print.loess_smooth}
\alias{plot.loess_smooth}
\title{Fit loess curves}
\usage{
loess_smooth(x, ...)

\method{loess_smooth}{formula}(formula, data, ...)

\method{loess_smooth}{default}(x, y = NULL, ..., conf = 0.95,
  n = 50L, sd = TRUE, fix_extremes = FALSE)

\method{print}{loess_smooth}(x, ...)

\method{plot}{loess_smooth}(x, y, ..., col.line = 1L,
  col.ci = col.line, alpha.ci = 0.5, xlab = NULL, ylab = NULL,
  ci = c("none", "lines", "band"))
}
\arguments{
\item{x, y}{the x- and y-axes for the plot}

\item{...}{additional arguments passed to or from other methods}

\item{formula}{a formula of the form \code{y ~ x} giving the x- and y-
axes for the plot}

\item{data}{a data frame (or list) from which the variables in
\code{formula} should be taken}

\item{conf}{the confidence interval to calculate}

\item{n}{the number of points at which to evaluate the smooth curve}

\item{sd}{logical; if \code{TRUE}, the standard deviation of the square
of the residuals are used to calculate the confidence intervals; otherwise,
the standard errors from \code{\link{predict.loess}} are used}

\item{fix_extremes}{logical; if \code{TRUE}, the extremes of the fitted
lines will run through \code{{x_1, y_1}} and \code{{x_n, y_n}}}

\item{col.line, col.ci}{colors for the smoothed curve and confidence
intervals (lines or band)}

\item{alpha.ci}{alpha transparency for the confidence band}

\item{ci}{type of confidence interval to plot, one of \code{"none"} for
no plotting, \code{"lines"} for lines only, or \code{"band"} for a
semi-transparent color band}
}
\description{
Add a smooth curve computed by \code{\link{loess}} to a scatter plot.
}
\examples{
plot(mpg ~ wt, mtcars)
lo <- loess_smooth(mpg ~ wt, mtcars)
lines(lo$x, lo$y)
lines(lo$x, lo$upper, lty = 2)
lines(lo$x, lo$lower, lty = 2)

plot(lo)
plot(lo, ci = 'lines')
plot(lo, ci = 'band')

}
\seealso{
\code{\link{loess.smooth}}
}
