% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown_report.R
\name{make_rmd_html_report}
\alias{make_rmd_html_report}
\title{Make an R Markdown report from a template}
\usage{
make_rmd_html_report(
  rmd_template,
  report_title = NULL,
  output_file_name,
  show_report = T,
  ...
)
}
\arguments{
\item{rmd_template}{Path to an R Markdown template.}

\item{report_title}{Title of the report.}

\item{show_report}{A logical value indicating whether or not to open the report after it's completion. Default is TRUE.}

\item{...}{Additional arguments required by the R Markdown template.}

\item{output_filepath}{Path to save output HTML document.}
}
\value{
Silently returns TRUE
}
\description{
Make an R Markdown report from a template
}
