% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_functions.R
\name{make_upset_plot}
\alias{make_upset_plot}
\title{Create an upset plot with a title and export}
\usage{
make_upset_plot(
  input_list,
  title = NULL,
  title_size = 30,
  filename = NULL,
  intersection_size_title_size = 2,
  intersection_size_tick_labels_size = 2,
  bar_numbers = 2,
  set_size_title_size = 2,
  set_size_tick_labels_size = 2,
  set_names_size = 2
)
}
\arguments{
\item{input_list}{A named list to use as input for the plot}

\item{title}{Title for the plot}

\item{title_size}{Size of the title in points. Default is 30.}

\item{filename}{Name of file to save plot to. If not provided, file not saved.}
}
\description{
Create an upset plot with a title and export
}
