% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.envir.tools.R
\name{mod.palette}
\alias{mod.palette}
\title{Modify color palette}
\usage{
mod.palette(palette, new.name, new.colors)
}
\arguments{
\item{palette}{palette to modify.}

\item{new.name}{new palette name.}

\item{new.colors}{colors to assign to palette.}
}
\description{
Modifies name and/or colors of a palette preset. Palettes provided with the package cannot be modified.
WARNING: changes to palettes are permanent and cannot be reverted.
}
\keyword{Color}
\keyword{Palette}
