% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-dataSettings.R
\name{modelSettings}
\alias{modelSettings}
\title{Server function for model settings}
\usage{
modelSettings(input, output, session, data, plotData)
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session}

\item{data}{(reactive) dataframe of the selected loaded file}

\item{plotData}{(reactive) list of selected saved data}
}
\description{
Backend for model settings module
}
