% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{no_axis_text}
\alias{no_axis_text}
\title{Theme with no axis text}
\usage{
no_axis_text()
}
\value{
A theme object that can be added to a ggplot object
}
\description{
A ggplot2 theme with no axis text
}
\examples{
\dontrun{
library(ggplot2)
ggplot(mtcars, aes(wt, mpg)) + 
  geom_point() + 
  no_axis_text()
}
}
