% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{no_gridlines}
\alias{no_gridlines}
\title{Theme with no grid lines}
\usage{
no_gridlines()
}
\value{
A theme object that can be added to a ggplot object
}
\description{
A ggplot2 theme with no grid lines
}
\examples{
\dontrun{
library(ggplot2)
ggplot(mtcars, aes(wt, mpg)) + 
  geom_point() + 
  no_gridlines()
}
}
