% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{pca}
\alias{pca}
\title{PCA plot}
\usage{
pca(data, group = rep(1L, ncol(data)), col = NULL, ..., centers = NULL,
  plot3d = FALSE, labels = FALSE, rpca = FALSE, legend = TRUE,
  args.legend = list())
}
\arguments{
\item{data}{an \code{r x c} matrix-like object with \code{r} features and
\code{c} samples}

\item{group}{a length \code{c} vector identifying the group of each sample}

\item{col}{a vector of colors for each group}

\item{...}{additional arguments passed to \code{\link{plot}} or further to
\code{\link{par}}}

\item{centers}{an integer value for the number of desired clusters, passed
to \code{\link[stats]{kmeans}}}

\item{plot3d}{logical; if \code{TRUE}, the first three dimensions are drawn
as a 3D scatterplot; see \code{\link[rgl]{plot3d}}}

\item{labels}{argument controlling labels for points; if \code{FALSE}, no
labels are drawn; if a numeric value in \code{(0,1)}, points outside
quantile are labelled; otherwise, a vector of labels for each point}

\item{rpca}{logical; if \code{TRUE}, \code{\link[rsvd]{rpca}} is used to
calculate the principal components}

\item{legend}{logical; if \code{TRUE}, a legend will be drawn}

\item{args.legend}{a \emph{named} list of \code{\link{legend}} arguments to
override defaults}
}
\description{
Wrapper function to plot a \code{\link{prcomp}} or \code{\link[rsvd]{rpca}}
object with optional \code{\link[stats]{kmeans}} groupings. Alternatively,
plot a 3D scatterplot of the first three principal components.
}
\examples{
## feature x sample matrix
dat <- t(unique(iris[, 1:4]))
grp <- unique(iris)$Species

pca(dat)
pca(dat, grp)
pca(dat, grp, rpca = TRUE)
pca(dat, grp, centers = 3, pch = 16)$km$cluster
pca(dat, grp, centers = 3, col = tcol(3:5, alpha = 0.5),
    pch = 16, cex = 1.5, las = 1L, bty = 'l', labels = 0.001,
    args.legend = list(x = 'topleft', title = 'Species'))

}
\seealso{
\code{\link{tsne}}
}
