% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_functions.R
\name{pca_ggplot}
\alias{pca_ggplot}
\title{Create a PCA plot using ggplot2}
\usage{
pca_ggplot(
  pca,
  PCs = c(1, 2),
  sample_subset = NULL,
  colour_groups = NULL,
  colours = NULL,
  shape_groups = NULL,
  labels = NULL,
  title = NULL,
  colour_legend_title = NULL,
  shape_legend_title = NULL,
  alpha = 1,
  save_image_path = NULL,
  width = 9,
  height = 9
)
}
\arguments{
\item{pca}{A prcomp or princomp object}

\item{PCs}{A numeric vector of length 2 specifying which PCs to use. Default are PCs 1 and 2}

\item{sample_subset}{An optional subset of samples to plot instead of all samples from the PCA}

\item{colour_groups}{An optional vector of the groups the observations belong to. Should be a character vector or a factor.}

\item{colours}{An optional vector of colours to use for the colour_groups if they are provided.}

\item{shape_groups}{An optional vector of the groups the observations belong to. Should be a character vector or a factor.}

\item{labels}{An optional vector of labels for the observations. If used then labels will be shown instead of points.}

\item{title}{A title for the plot. Default is no title.}

\item{colour_legend_title}{The title of the colour legend for the plot. Default is no title.}

\item{shape_legend_title}{The title of the shape legend for the plot. Default is no title.}

\item{alpha}{Alpha value for the points}

\item{save_image_path}{An optional filepath where the image will be saved.}

\item{width}{If the image is saved, the width of the saved image. Default is 9 inches}

\item{height}{If the image is saved, the height of the saved image. Default is 9 inches}
}
\value{
A ggplot object
}
\description{
Create a PCA plot using ggplot2
}
