% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_functions.R
\name{pca_var_percent}
\alias{pca_var_percent}
\title{Calculate percentage of variation each PC in a PCA accounts for}
\usage{
pca_var_percent(pca)
}
\arguments{
\item{pca}{A prcomp or princomp object}
}
\value{
A numeric vector with the percentage of variation explained by each PC
}
\description{
Calculate percentage of variation each PC in a PCA accounts for
}
