% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_functions.R
\name{pdf_save}
\alias{pdf_save}
\title{Save a list of plots to a PDF file}
\usage{
pdf_save(plotlist, nrows = 1, ncols = 1, filename, width = 16, height = 9)
}
\arguments{
\item{plotlist}{A list of ggplot objects}

\item{nrows}{Number of rows in each sheet}

\item{ncols}{Number of columns in each sheet}

\item{filename}{Name of output file}

\item{width}{Width of each page. Default is 16.}

\item{height}{Heigth of each page. Default is 9.}
}
\value{
Nothing
}
\description{
Save a list of plots to a PDF file
}
