% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_chart.R
\name{plot_chart}
\alias{plot_chart}
\title{Create an Observable Plot chart}
\usage{
plot_chart(
  spec = list(),
  data = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{spec}{A list containing the plot specification including marks and options}

\item{data}{Optional data frame to use with the plot}

\item{width}{Width of the plot in pixels (NULL for automatic sizing)}

\item{height}{Height of the plot in pixels (NULL for automatic sizing)}

\item{elementId}{HTML element ID for the widget}
}
\value{
An htmlwidget object
}
\description{
Create an interactive data visualization using Observable Plot
}
\examples{
\dontrun{
# Simple scatter plot
plot_chart(
  spec = list(
    marks = list(
      Plot.dot(cars, list(x = "speed", y = "dist"))
    )
  ),
  data = cars
)
}

}
