% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_confint.R
\name{plot_lm_confint}
\alias{plot_lm_confint}
\title{Produces a lattice plot with an array of confidence intervals plots.}
\usage{
plot_lm_confint(formula, data, cex = 1, ...)
}
\arguments{
\item{formula}{a formula object of the form \code{y ~ x | c}.
\code{y} is a response, \code{x} is a covariate, and \code{c}
is a conditioning variable for which to generate plot panels.}

\item{data}{a data frame.}

\item{cex}{relative character size of labels.}

\item{...}{further arguments.}
}
\value{
A \code{\link[lattice]{trellis.object}}.
}
\description{
Plots 95% confidence interval on within-subjects intercept and slope.
Orders subjects by increasing intercept to check for relationships
between slopes and intercepts.
}
\details{
See:  Bates, D. M. (2010). "lme4: Mixed-effects modeling with R." 73.
}
\examples{
\dontrun{
plot_lm_confint(errors ~ mode | PID, data = data2, cex = 0.8)
}
}
\seealso{
\code{\link[lattice]{dotplot}}
}
\author{
Nathan Lucas
}
