% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_scatter.R
\name{plot_lm_scatter}
\alias{plot_lm_scatter}
\title{Produces a lattice plot with an array of scatter plots in panels.
Each panel includes a reference line obtained by simple linear regression.
The aspect ratio of the panels (width-to-height) is adjusted so a typical
reference line is ~45 degrees, which makes it easier to detect systematic
changes in slopes.
The panels are ordered by increasing intercept of simple linear regression,
not by PID (which is a random order).  If the slopes and the intercepts are
highly correlated we should see a pattern across the panels in the slopes.}
\usage{
plot_lm_scatter(formula, data, cex = 1, jitter = 0.5, ...)
}
\arguments{
\item{formula}{a formula object of the form \code{y ~ x | c}.
\code{x} and \code{y} are x- and y-axis variable names respectively,
and \code{c} is a conditioning variable for which to generate plot panels.}

\item{data}{a data frame.}

\item{cex}{relative character size of labels.}

\item{jitter}{amount of X-axis jittering, or \code{0} for no jitter.}

\item{...}{further arguments.}
}
\value{
A \code{\link[lattice]{trellis.object}}.
}
\description{
Produces a lattice plot with an array of scatter plots in panels.
Each panel includes a reference line obtained by simple linear regression.
The aspect ratio of the panels (width-to-height) is adjusted so a typical
reference line is ~45 degrees, which makes it easier to detect systematic
changes in slopes.
The panels are ordered by increasing intercept of simple linear regression,
not by PID (which is a random order).  If the slopes and the intercepts are
highly correlated we should see a pattern across the panels in the slopes.
}
\examples{
\dontrun{
plot_lm_scatter(y ~ x | group, data = dat, cex = 0.8)
}
}
\seealso{
\code{\link[lattice]{xyplot}}
}
\author{
Nathan Lucas
}
