% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marks.R
\name{plot_marks}
\alias{plot_marks}
\alias{plot_dot}
\alias{plot_line}
\alias{plot_barY}
\alias{plot_barX}
\alias{plot_area}
\alias{plot_rect}
\alias{plot_text}
\alias{plot_cell}
\alias{plot_ruleX}
\alias{plot_ruleY}
\alias{plot_link}
\alias{plot_frame}
\alias{plot_axisX}
\alias{plot_axisY}
\alias{plot_gridX}
\alias{plot_gridY}
\title{Create Plot marks for use in plot_chart}
\usage{
plot_dot(data = NULL, options = list())

plot_line(data = NULL, options = list())

plot_barY(data = NULL, options = list())

plot_barX(data = NULL, options = list())

plot_area(data = NULL, options = list())

plot_rect(data = NULL, options = list())

plot_text(data = NULL, options = list())

plot_cell(data = NULL, options = list())

plot_ruleX(data = NULL, options = list())

plot_ruleY(data = NULL, options = list())

plot_link(data = NULL, options = list())

plot_frame(options = list())

plot_axisX(options = list())

plot_axisY(options = list())

plot_gridX(options = list())

plot_gridY(options = list())
}
\arguments{
\item{data}{Data frame or NULL to use data from plot_chart}

\item{options}{List of options for the mark (x, y, fill, r, etc.)}
}
\description{
These functions create mark specifications that can be used with plot_chart.
They mirror the JavaScript Plot.* mark functions.
}
\section{Functions}{
\itemize{
\item \code{plot_dot()}: Create a dot mark

\item \code{plot_line()}: Create a line mark

\item \code{plot_barY()}: Create a bar mark for vertical bars

\item \code{plot_barX()}: Create a bar mark for horizontal bars

\item \code{plot_area()}: Create an area mark

\item \code{plot_rect()}: Create a rect mark

\item \code{plot_text()}: Create a text mark

\item \code{plot_cell()}: Create a cell mark

\item \code{plot_ruleX()}: Create a rule mark for vertical lines

\item \code{plot_ruleY()}: Create a rule mark for horizontal lines

\item \code{plot_link()}: Create a link mark

\item \code{plot_frame()}: Create a frame mark

\item \code{plot_axisX()}: Create an axisX mark

\item \code{plot_axisY()}: Create an axisY mark

\item \code{plot_gridX()}: Create a gridX mark

\item \code{plot_gridY()}: Create a gridY mark

}}
