% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_chart.R
\name{plot_output}
\alias{plot_output}
\title{Shiny output binding for plotR}
\usage{
plot_output(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{Output variable to read from}

\item{width}{Width of the plot}

\item{height}{Height of the plot}
}
\description{
Shiny output binding for plotR
}
