% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_png.R
\name{plot_png}
\alias{plot_png}
\title{Save Plot to PNG}
\usage{
plot_png(x, file, prt = TRUE, ...)
}
\arguments{
\item{x}{a plot to save.}

\item{file}{name of output file without extension, up to 507 characters.}

\item{prt}{\code{TRUE} to also print plot.}

\item{...}{more parameters passed to \code{png()}.}
}
\value{
Returns the file name of the output PNG image.
}
\description{
Saves a plot as a PNG file and optionally prints the plot.
}
\examples{
dir.create("plot", showWarnings = FALSE, recursive = TRUE)
plot_png(plot(sin, -pi, pi), file = "plot/sine")
}
\author{
Nathan Lucas
}
