% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotr-package.R
\docType{package}
\name{plotr-package}
\alias{plotr}
\alias{plotr-package}
\title{plotr: Make Dealing with Plots a Little Bit Easier}
\description{
Complicated plot functions ready to use with a single line of code.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://danielvartan.github.io/plotr/}
  \item \url{https://github.com/danielvartan/plotr/}
  \item Report bugs at \url{https://github.com/danielvartan/plotr/issues/}
}

}
\author{
\strong{Maintainer}: Daniel Vartanian \email{danvartan@gmail.com} (\href{https://orcid.org/0000-0001-7782-759X}{ORCID}) [conceptor, copyright holder]

}
\keyword{internal}
