% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotr.R
\docType{package}
\name{plotr}
\alias{plotr}
\alias{plotr-package}
\title{Plotting utilities for R}
\description{
Utility functions to make plotting easier and more consistent.
}
\author{
Nathan Lucas
}
