% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zzz.R
\name{plotr_utils}
\alias{\%inside\%}
\alias{\%ni\%}
\alias{\%||\%}
\alias{inside}
\alias{notin}
\alias{oror}
\alias{plotr_utils}
\alias{tcol}
\title{plotr utils}
\usage{
x \%ni\% table

x \%inside\% interval

a \%||\% b

tcol(color, trans = 255)
}
\arguments{
\item{x}{vector or \code{NULL}; the values to be matched}

\item{table}{vector or \code{NULL}; the values to be matched against}

\item{interval}{numeric vector of length two representing the interval}

\item{a,b}{raw, logical, "number-like" vectors or objects}

\item{color}{vector of color names (or hexadecimal)}

\item{trans}{transparency defined as an integer in the range
\code{[0, 255]} where \code{0} is fully transparent and \code{255} is fully
visible}
}
\description{
Some utilities for \code{plotr}.
}
\details{
\code{\%ni\%} is the negation of \code{\link[base]{\%in\%}}.

\code{\%inside\%} returns a logical vector indicating if \code{x} is inside
the \code{interval} (inclusive).

\code{\%||\%} is useful for a function, \code{f}, that may return a value
or \code{NULL}, but if \code{NULL} is the result of \code{f}, it is
desirable to return some other default value without errors.

\code{tcol} adds transparency to colors (vectorized).
}
\examples{
\dontrun{
1:5 \%ni\% 3:5

c(0,4) \%inside\% c(0, 4)
-5:5 \%inside\% c(0,5)
-5:5 \%inside\% c(5,0)

f <- function(x0 = TRUE) NULL || x0
f() # error
f <- function(x0 = TRUE) NULL \%||\% x0
f() # TRUE

cols <- c('red','green','blue')

# a normal plot
plot(rnorm(100), col = tcol(cols), pch = 16, cex = 4)

# more transparent
plot(rnorm(100), col = tcol(cols, 100), pch = 16, cex = 4)

# hexadecimal colors also work
cols <- c('#FF0000','#00FF00','#0000FF')
plot(rnorm(100), col = tcol(cols, c(50, 100, 255)), pch= 16, cex = 4)
}
}
\seealso{
\code{\link{\%in\%}}, \code{\link{||}}
}

