% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_extra.R
\name{polygon2}
\alias{polygon2}
\title{polygon2}
\usage{
polygon2(x, y = NULL, radius, sides = 6, srt = 0, density = NULL,
  angle = 45, border = NULL, col = NA, lty = par("lty"),
  fillOddEven = FALSE, ...)
}
\arguments{
\item{x, y}{x- and y-coordinate vectors of polygon centroids}

\item{radius}{distance from centroid to each vertex}

\item{sides}{number of sides for polygons; since only regular polygons may
be plotted, possible values are \code{(1:360)[360L \%\% 1:360 == 0L]}}

\item{srt}{rotation in degrees for polygon}

\item{density}{density of shading lines in lines per inch}

\item{angle}{slope of shading lines, given as an angle in degrees (counter-
clockwise)}

\item{border}{polygon border color}

\item{col}{polygon fill color}

\item{lty}{line type to be used for border and/or shading lines}

\item{fillOddEven}{logical controlling polygon shading mode}

\item{...}{additional graphical parameters such as \code{xpd}, \code{lend},
\code{ljoin}, and \code{lmitre} passed to \code{\link{par}}}
}
\description{
Draw \emph{regular} \code{\link{polygon}}s as plotting characters. This
function calculates the \code{{x,y}} coordinates for given centroids and
radii and passes each pair of vectors to \code{\link{polygon}} to draw.
Note that all arguments except \code{x}, \code{y}, \code{radius},
\code{sides}, and \code{srt} are passed directly to \code{\link{polygon}},
so see \code{?polygon} for additional details.
}
\examples{
## sides can be any of
(sides <- (1:360)[360L \%\% 1:360 == 0L])
plot.new()
for (ii in sides)
  polygon2(0.5, 0.5, sides = ii)  ## okay
# polygon2(0.5, 0.5, sides = 7)   ## error

x <- mtcars$mpg
y <- mtcars$wt
plot(x, y, type = 'n', asp = 1)
polygon2(x, y, density = 30, angle = 90, col = mtcars$gear)
polygon2(x, y + 1, srt = 30, lty = 'dotted', col = 'transparent')
polygon2(x, y + -2, radius = .5, sides = 5)

}
