% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-predictAdditionalData.R
\name{predictAdditionalData}
\alias{predictAdditionalData}
\title{Predict Estimations for More X Input}
\usage{
predictAdditionalData(moreMean, moreSD, plotValues, moreNSample = NULL)
}
\arguments{
\item{moreMean}{numeric mean input value}

\item{moreSD}{numeric sd for input value}

\item{plotValues}{list of plotValues from a savedPlots element}

\item{moreNSample}{(numeric) number of samples for the input}
}
\description{
Predict Estimations for More X Input
}
