% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_special.R
\name{prettybars2}
\alias{prettybars2}
\title{prettybars2}
\usage{
prettybars2(x, lab.y = colnames(x), n.y = ncol(x),
  lab.group = rownames(x), n.group = nrow(x),
  col.group = tcol(seq.int(n.group), alpha = 0.5), col.line = "skyblue3",
  extra.margin = 0, col.bg = "snow", cex.y = 0.7, notext = FALSE,
  title = paste0("prettybar of ", m$x), sub = paste0("     n = ", sum(x)),
  note = NULL, subnote = "subnote: here is a subnote", legend = 1L, ...)
}
\arguments{
\item{x}{a table of data, typically from \code{\link{table}} or 
\code{\link{xtabs}}}

\item{lab.y}{y-axis labels}

\item{n.y}{number of y-groups (calculated from \code{x})}

\item{lab.group}{group labels}

\item{n.group}{number of groups (calculated from \code{x})}

\item{col.group}{color for each group level}

\item{col.line}{color of origin line}

\item{extra.margin}{extra left spacing for long \code{lab.y} labels}

\item{col.bg}{background color}

\item{cex.y}{size of \code{lab.y}}

\item{notext}{logical; if \code{TRUE}, suppresses all plot text, labels, and
axes for user input}

\item{title, sub}{main and sub titles}

\item{note, subnote}{optional note (top right) or subnote (bottom right)}

\item{legend}{integer (0, 1, 2); \code{0} for no legend; \code{1} for a
legend with pre-formatting; or \code{2} for a default \code{\link{legend}}}

\item{...}{additional graphical parameters passed to \code{\link{par}}}
}
\description{
A barplot.
}
\examples{
set.seed(1)
f <- function(...) sample(1:5, 100, replace = TRUE, prob = c(...))
dat <- data.frame(q1 = f(.1, .2, .3, .3, .1),
                  q2 = f(.1, .4, .1, .3, .1),
                  q3 = f(.1, .2, .3, .3, .1),
                  q4 = f(.1, .1, .3, .3, .1),
                  q5 = f(.2, .1, .2, .3, .2),
                  q6 = f(.1, .3, .3, .2, .1),
                  q7 = f(.1, .4, .1, .1, .3))
dat <- stack(dat)
dat <- within(dat, {
  values <- factor(values, levels = 1:5, labels = c('NA','SA','A','D','SD'))
})

tdat <- table(dat)
cols <- c(
  grey(.9),
  rawr::tcol(c('lightblue', 'lightblue', 'magenta1', 'magenta1'),
             alpha = c(.75, .25, .25, .75))
)
               
## compare:
barplot(tdat, horiz = TRUE, las = 1, col = cols, border = NA)
prettybars2(tdat, lab.y = paste('Question #', 1:7), extra.margin = 3,
            col.group = cols)

## using bibar
par(mar = c(5, 8, 4, 2))
at <- bibar(t(tdat), 3:2, 4:5, 1, col = cols,
            xlim = c(-80, 50), axes = FALSE)

axis(1, pretty(c(-80, 50)), abs(pretty(c(-80, 50))))
abline(v = 0, col = 'lightblue', lwd = 4)
text(-85, at, sprintf('Question # \%s', 1:7), xpd = NA, adj = 1)

legend('topleft', inset = c(0, -0.2), xpd = NA, fill = cols[2:3],
       legend = rownames(tdat)[2:3], horiz = TRUE, bty = 'n')
legend('topright', inset = c(0, -0.2), xpd = NA, fill = cols[4:5],
       legend = rownames(tdat)[4:5], horiz = TRUE, bty = 'n')

}
\seealso{
\code{\link{prettybars}}; \code{\link{bibar}}
}
