% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plots.R
\name{prettypie}
\alias{prettypie}
\title{prettypie}
\usage{
prettypie(dat, file, dev = "pdf", width = 15, height = 11, main, sub,
  note)
}
\arguments{
\item{dat}{data}

\item{file}{path to output file}

\item{dev}{device; default is \code{\link{pdf}}}

\item{width}{width of \code{dev}}

\item{height}{height of \code{dev}}

\item{main}{overall title for plot}

\item{sub}{sub-title for plot}

\item{note}{note for plot}
}
\description{
A pie chart
}
\examples{
\dontrun{
browsers <- source(system.file('source','browsers.r', package = 'plotr'))$value
browsers <- within(browsers, total <- ave(share, browser, FUN = sum))
browsers <- browsers[c('browser','version','share','total')]

prettypie(dat = browsers,
          file = './donuts.pdf',
          main = 'Browser market share, April 2011',
          sub = 'stackoverflow.com:::maryam',
          note = '/questions/26748069/ggplot2-pie-and-donut-chart-on-same-plot')


disease <- source(system.file('source','disease.r', package = 'plotr'))$value
disease <- disease[c('Disease','Acosts60','Pcosts60','Patients60','Total60')]

prettypie(dat = disease,
          file = './cost.pdf',
          main = 'Cost of getting sick',
          sub = 'Inside: Personal Costs, outside: Insurer Costs',
          note = 'visualization.geblogs.com/visualization/health_costs/')
}
}

