% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.envir.R
\name{print.single}
\alias{print.single}
\title{Print color palette}
\usage{
\method{print}{single}(p, palette)
}
\arguments{
\item{p}{palette list}

\item{palette}{name of palette to use}
}
\description{
Workhorse function to print color palettes. Not available to users.
}
\keyword{Workhorse}
