% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2.R
\name{propfall}
\alias{propfall}
\title{propfall}
\usage{
propfall(data, group, order, group.order = unique(group), col = NULL,
  ...)
}
\arguments{
\item{data}{a data frame (or object to be coerced)}

\item{group}{a vector with length \code{nrow(data)} defining the group to
which each row of \code{data} belongs}

\item{order}{a vector of column names of \code{data} defining how numeric
values should be sorted}

\item{group.order}{an optional vector of unique values of \code{group}
giving the desired order for groups}

\item{col}{a vector of colors for each of \code{order}, recycled as needed}

\item{...}{additional parameters passed to \code{\link{barplot}} or
graphical parameters passed to \code{\link{par}}}
}
\value{
A list of length two with the result of the call to \code{\link{barplot}}
(\code{bp}) giving the x-axis positions for each bar and the \code{o}rder
that each observation has been sorted in the barplot.
}
\description{
A barplot of proportions.
}
\examples{
dat <- within(mtcars, {
  disp <- disp / 10
  wt <- wt * 10
})

vars <- c('mpg', 'disp', 'wt')
dat[, vars] <- t(apply(dat[, vars], 1L, function(x) x / sum(x)))

dat <- dat[, vars]
dat$group <- colnames(dat)[max.col(dat)]

cols <- c('grey', 'lightpink', 'indianred1', 'indianred3')
propfall(dat[, 1:3])
propfall(dat[, 1:3], group = dat$group, col = cols)

## use the return value to add labels or identify observations
bp <- propfall(
  dat[, 1:3], col = cols, group = dat$group, order = c('disp', 'wt', 'mpg'),
  group.order = c('mpg', 'wt', 'disp')
)
text(bp$bp, 0, labels = rownames(dat)[bp$o], srt = 90, col = 0, adj = 0)

}
\seealso{
\code{\link[rawr]{waterfall}}
}
