% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_extra.R
\name{ptlocator}
\alias{ptlocator}
\title{Point locator}
\usage{
ptlocator(
  n = 512L,
  x,
  y,
  col = adjustcolor("red", alpha.f = 0.5),
  pch = 16L,
  ...
)
}
\arguments{
\item{n}{the maximum number of points to locate}

\item{x, y}{x- and y-coordinates}

\item{col}{color for points}

\item{pch}{plotting character}

\item{...}{additional graphical arguments passed to \code{\link{points}}}
}
\value{
Returns (invisibly) the indices of selected points.
}
\description{
Interactively select points on an existing figure and annotated with new
plotting characters.
}
\examples{
\dontrun{
set.seed(1)
n <- 200
x <- sort(runif(n, 0, 10 * pi))
y <- sin(x) + rnorm(n, 0, .2)

plot(x, y, cex = 2)
p <- ptlocator(10, x, y, cex = 2)
cbind(x = x[p], y = y[p])
}

}
\references{
\url{http://www.r-bloggers.com/data-point-locator-function/}
}
