% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-predictAdditionalData.R
\name{rdensity}
\alias{rdensity}
\title{Generate n random numbers from density() object}
\usage{
rdensity(n, x, probs, bw)
}
\arguments{
\item{n}{The total random numbers to generate}

\item{x}{numeric vector}

\item{probs}{probabilities for x}

\item{bw}{bandwidth used in density()}
}
\description{
Generate n random numbers from density() object
}
