% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-downAndUploadModel.R
\name{removeModelOutputs}
\alias{removeModelOutputs}
\title{Remove Model Outputs}
\usage{
removeModelOutputs(models)
}
\arguments{
\item{models}{list of model objects to be saved}
}
\description{
Remove Model Outputs
}
