% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_chart.R
\name{render_plot}
\alias{render_plot}
\title{Shiny render binding for plotR}
\usage{
render_plot(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An expression that generates a plotR chart}

\item{env}{The environment in which to evaluate expr}

\item{quoted}{Is expr a quoted expression}
}
\description{
Shiny render binding for plotR
}
