% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.envir.tools.R
\name{save.palette}
\alias{save.palette}
\title{Save palette as preset}
\usage{
save.palette(palette, col)
}
\arguments{
\item{palette}{name of palette}

\item{col}{vector of 6-digit HEX values}
}
\description{
Saves a vector of 6-digit HEX values as a palette.
}
\keyword{Color}
\keyword{Palette}
