% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{scattergram}
\alias{scattergram}
\title{Scattergram}
\usage{
scattergram(
  x,
  y,
  g = NULL,
  col = NULL,
  ...,
  add = FALSE,
  xlim = NULL,
  ylim = NULL,
  axs = "i",
  heights = 0.25,
  main = NULL,
  legend = TRUE,
  args.legend = list()
)
}
\arguments{
\item{x, y, g}{the x-, y-, and group variables}

\item{col}{a vector of colors or integers for each unique value of \code{g}}

\item{...}{additional arguments passed to \code{plot} (e.g., \code{xlim},
\code{panel.first}) or further to \code{par}}

\item{add}{logical; if \code{TRUE}, plots will be added to an existing
or multi-panel figure}

\item{xlim, ylim}{the x- and y-axis limits}

\item{axs}{the style(s) of axis interval calculation; see \code{\link{par}}}

\item{heights}{for \code{add = TRUE}, the relative height of each density
panel compared to the figure region}

\item{main}{the main figure title}

\item{legend}{logical; if \code{TRUE}, a \code{group} legend is drawn}

\item{args.legend}{a \emph{named} list of \code{\link{legend}} arguments to
override defaults}
}
\description{
A scattergram plot with density curves along axes.
}
\examples{
with(mtcars, {
  scattergram(mpg, wt)
})

with(mtcars, {
  scattergram(mpg, wt, cyl, col = rainbow(3), pch = 16)
})

with(mtcars, {
  scattergram(
    mpg, wt, c('Four', 'Six', 'Eight')[cyl / 2 - 1],
    xlab = 'MPG', ylab = 'Weight', las = 1,
    panel.first = abline(lm(wt ~ mpg), col = 5, lwd = 3),
    panel.last = abline(v = c(20, 25))
  )
})


## to add multiple figures on a plot, use add = TRUE
sp <- split(mtcars, mtcars$am)
op <- par(mfrow = 1:2, mar = c(5, 5, 5, 4), oma = c(0, 0, 0, 2))
lapply(sp, function(d) {
  scattergram(d$mpg, d$wt, d$cyl, col = 1:3, add = TRUE)
})
par(op)

}
