% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scattermat.R
\name{scattermat}
\alias{scattermat}
\title{Scatterplot Matrix}
\usage{
scattermat(x, data, cor.method = "pearson", bold.limit = 0.7, pch = 1,
  cex.axis = 1, cex.cor = 1, cex.labels = 1, cex.main = 1, ...)
}
\arguments{
\item{x}{a formula describing the variables to plotted.}

\item{data}{a data frame.}

\item{cor.method}{correlation method.}

\item{bold.limit}{coefficients >= this threshold are show in bold text.}

\item{pch}{plot point symbol (default is a circle).}

\item{cex.axis}{relative character size of axis labels.}

\item{cex.cor}{relative character size of correlation coefficient text.}

\item{cex.labels}{relative character size of ... .}

\item{cex.main}{relative character size of plot title.}

\item{...}{more parameters passed to ... .}
}
\value{
Returns results of the correlation tests.
}
\description{
Creates a scatter plot matrix.
}
\author{
Nathan Lucas
}
