% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_functions.R
\name{scree_plot}
\alias{scree_plot}
\title{Create a scree plot showing the percentage of variation each PC explains using ggplot2}
\usage{
scree_plot(
  pca,
  number_of_pcs = 10,
  title = NULL,
  xlab = NULL,
  ylab = NULL,
  rotate_x_tick_labels = F,
  colour_palette = "Oranges"
)
}
\arguments{
\item{pca}{A prcomp or princomp object}

\item{number_of_pcs}{Number of PCs to display in the plot. Default is 10 or the total number of PCs, whichever is smaller.}

\item{title}{Title for the plot. Default is "Percentage of Variance Explained by Top \code{number_of_pcs} PCs".}

\item{xlab}{Label for the x axis. Default is "PC".}

\item{ylab}{Label for the y axis. Default is "Percentage of Variance Explained".}

\item{colour_palette}{Either of vector of colours with length equal to \code{number_of_pcs} or the name of a palette from RColorBrewer. Default is the "Oranges" palette.}

\item{rotate_x_axis_labels}{Whether to rotate the x labels by 90 degrees. Default is False}
}
\value{
A ggplot object
}
\description{
Create a scree plot showing the percentage of variation each PC explains using ggplot2
}
