% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-prepareData.R
\name{selectDataWrapper}
\alias{selectDataWrapper}
\title{Select Data Wrapper}
\usage{
selectDataWrapper(plotValues, activeFile, activeFileData, dataSelection)
}
\arguments{
\item{plotValues}{(list) list with all values required for the plot}

\item{activeFile}{(character) name of the selected file}

\item{activeFileData}{(data.frame) content of the selected file}

\item{dataSelection}{(list) list of reactive selected columns of the current file}
}
\description{
Select Data Wrapper
}
