% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplots.R
\name{set_panel_size}
\alias{set_panel_size}
\title{Set panel size}
\usage{
set_panel_size(
  p = NULL,
  file = NULL,
  margin = unit(1, "cm"),
  width = unit(4, "cm"),
  height = unit(4, "cm")
)
}
\arguments{
\item{p}{a \code{\link[ggplot2]{ggplot}} object}

\item{file}{optional file name if result should be saved to file, passed
to \code{\link[ggplot2]{ggsave}}}

\item{margin}{a \code{\link[grid]{unit}} object giving the plot margins}

\item{width, height}{a \code{\link[grid]{unit}} object giving height and
width for each plot}
}
\description{
Use strict size for panels when number of panels in rows/columns are not
equal.
}
\examples{
library('ggplot2')
library('gridExtra')
p1 <- ggplot(mtcars, aes(mpg, wt)) + geom_point() + facet_wrap(~ vs)
p2 <- ggplot(mtcars, aes(mpg, wt)) + geom_point() + facet_wrap(~ gear)
grid.arrange(p1, p2)

g1 <- set_panel_size(p1)
g2 <- set_panel_size(p2)
grid.arrange(g1, g2)

}
\references{
\href{http://stackoverflow.com/questions/32580946/setting-absolute-size-of-
facets-in-ggplot2}{SO question}
}
\author{
Baptiste Auguie
}
