% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2.R
\name{shist}
\alias{shist}
\title{shist}
\usage{
shist(x, ..., col = grey.colors(length(x) + 1L), loess = TRUE,
  total = TRUE, xlim = NULL, ylim = NULL, heights = NULL,
  heights.main = 0.5, reset_par = TRUE)
}
\arguments{
\item{x}{a list of numeric vectors for which the histogram is desired}

\item{...}{additional parameters passed to \code{\link{hist}} or graphical
parameters passed to \code{\link{par}}}

\item{col}{a vector of colors for each histogram}

\item{loess}{logical; if \code{TRUE}, a \code{\link{loess}} curve is
fit to each histogram}

\item{total}{logical; if \code{TRUE}, an aggregate histogram of \code{x}
is plotted first, and each element of \code{x} is plotted below}

\item{xlim, ylim}{the x- and y-axis limits (if given, applied to all plots)}

\item{heights}{an optional vector of values, usually in \code{(0,1)}, for
the heights of each plot on the device; see \code{\link{layout}}}

\item{heights.main}{if \code{total = TRUE}, the height of the aggregate
histogram plot relative to the device, usually in \code{(0,1)}}

\item{reset_par}{logical; if \code{TRUE}, resets par settings to state
before function call; setting \code{reset_par = FALSE} is useful for adding
to a plot}
}
\value{
A list of length \code{length(x)} containing the return value of
\code{\link{hist}} for each element of \code{x}.
}
\description{
A stacked histogram with density curve.
}
\examples{
set.seed(1)
x <- lapply(sample(1:10, 4), rpois, n = 500)

shist(x)
shist(x, heights.main = 0.75)
shist(x, heights = c(5,1,1,1,3))
shist(x, col = rainbow(5), total = FALSE,
      ylim = c(0, 150), las = 1L, breaks = 10)

}
