% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_functions.R
\name{sig_sym}
\alias{sig_sym}
\title{Encode p-values with symbols}
\usage{
sig_sym(p_values, cutpoints = c(0, 0.001, 0.01, 0.05, 1), symbol = "*")
}
\arguments{
\item{p_values}{A vector of p-values}

\item{cutpoints}{A vector of cutpoints for significance. Default is c(0, 0.001, 0.01, 0.05, 1)}
}
\value{
A vector of symbols representing p-values
}
\description{
Encode p-values with symbols
}
