% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skip_label.R
\name{skip_label}
\alias{skip_label}
\title{Create intervals of missing labels for \code{ggplot2} charts}
\usage{
skip_label(x, type = "even")
}
\arguments{
\item{x}{A \code{\link[base:character]{character}} vector of labels}

\item{type}{(Optional) A \code{\link[base:character]{character}} string with the
type of labels to remove. Possible values are \code{"even"}, \code{"odd"} and
\code{"one"} (Default: \code{"even"}).}
}
\value{
A \code{\link[base:character]{character}} vector of labels with some
them missing (\code{""}).
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

\code{skip_label()} is a utility function that creates intervals of missing
labels for \href{https://ggplot2.tidyverse.org/}{\code{ggplot2}} charts. This is
useful when you have a large number of labels in a axis and you want to
remove some of them.

Use this function in the \code{label} argument of functions like
\code{\link[ggplot2:scale_continuous]{scale_x_continuous()}} or
\code{\link[ggplot2:scale_discrete]{scale_y_discrete()}}.
}
\examples{
skip_label(x = 0:5, type = "even")
#> [1] "0" ""  "2" ""  "4" "" # Expected

skip_label(x = 0:5, type = "odd")
#> [1] ""  "1" ""  "3" ""  "5" # Expected

skip_label(x = 0:5, type = "one")
#> [1] "0" "1" ""  "3" "4" "" # Expected
}
\concept{utility functions}
