% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-combinedInputModule.R
\name{sliderAndNumericRangeServer}
\alias{sliderAndNumericRangeServer}
\title{Slider And Input Server}
\usage{
sliderAndNumericRangeServer(id, value, min, max)
}
\arguments{
\item{id}{id of module}

\item{value}{(numeric) default value}

\item{min}{(numeric) minumum}

\item{max}{(numeric) maximum}
}
\description{
Server function of the Slider And Input module
}
