% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-combinedInputModule.R
\name{sliderAndNumericRangeUI}
\alias{sliderAndNumericRangeUI}
\title{Slider And Input UI}
\usage{
sliderAndNumericRangeUI(id, label, min, max, value, dragRange)
}
\arguments{
\item{id}{id of module}

\item{label}{label}

\item{min}{(numeric) minumum}

\item{max}{(numeric) maximum}

\item{value}{(numeric) default value}

\item{dragRange}{(logical) dragRange}
}
\description{
UI of the Slider And Input module
}
