% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_extra.R
\name{subplot}
\alias{subplot}
\title{Subplot}
\usage{
subplot(
  expr,
  x,
  y = NULL,
  log = NULL,
  size = c(1, 1),
  vadj = 0.5,
  hadj = 0.5,
  inset = c(0, 0),
  type = c("plt", "fig"),
  pars = NULL
)
}
\arguments{
\item{expr}{an expression defining the plot to be embedded}

\item{x, y}{coordinates of the new plot given by a keyword (see details)
or values accepted by \code{\link{xy.coords}}; if missing, the user
will be prompted to select the bottom-left and top-right coordinates for
the plotting region}

\item{log}{character, \code{"x"}, \code{"y"} or both, as for
\code{\link{plot}}; sets negative values to \code{NA} and gives a warning}

\item{size}{the size of the plot in inches if \code{x} and \code{y} are
length 1}

\item{vadj}{vertical adjustment of the plot when \code{y} is a scalar; the
default is to center vertically; 0 means place the bottom of the plot at
\code{y}, and 1 places the top of the plot at \code{y}}

\item{hadj}{horizontal adjustment of the plot when \code{x} is a scalar,
the default is to center horizontally; 0 means place the left edge of the
plot at \code{x}, and 1 means place the right edge of the plot at \code{x}}

\item{inset}{1 or 2 numbers representing the proportion of the plot to
inset the subplot from edges when \code{x} is a character string; the first
element is the horizontal inset, and the second is the vertical inset}

\item{type}{character string of \code{"plt"} or \code{"fig"}; if "plt"
then the plotting region is defined by \code{x}, \code{y}, and \code{size}
with axes, etc. outside the box; if "fig" then all annotations are also
inside the box}

\item{pars}{a list of parameters to be passed to \code{\link{par}} before
running \code{expr}}
}
\value{
The list of graphical parameters used when drawing \code{expr}, useful for
adding an additional subplot after \code{subplot} has finished.
}
\description{
Embed a subplot in an existing plot at specified coordinates.
}
\details{
The coordinates for \code{expr} may be given by a keyword passed to
\code{x} (similar to \code{\link{legend}}). \code{x} will be grepped for
"right", "left", "top", and/or "bottom", and the coordinates for
\code{expr} will be automatically calculated; see examples.

If \code{x} and \code{y} are numeric and length 1, these will be used as
the center of the plotting region; if length 2, these will be used as the
minimum and maximum coordinates of the plotting region (similarly if
\code{x} is missing and the user is prompted by \code{\link{locator}} to
give the coordinates interactively).
}
\examples{
plot(1)
subplot(plot(density(rnorm(100)), ann = FALSE), 1.2, 1.2)
subplot(plot(density(rnorm(100)), ann = FALSE, axes = FALSE), 'bottomright')


\dontrun{
## augment a map
library('maptools')
data(state.vbm)

plot(state.vbm, fg = NULL)
tmp <- cbind(state.vbm$center_x, state.vbm$center_y)

for (i in 1:50) {
  tmp2 <- as.matrix(USArrests[i, c(1, 4)])
  tmp3 <- max(USArrests[, c(1, 4)])
  subplot({
    barplot(matrix(tmp2), ylim = c(0, tmp3), yaxt = 'n', col = 1:2,
            beside = TRUE, border = 1:2)
   }, x = tmp[i, 1], y = tmp[i, 2], size = c(.1, .1))
}
legend('bottomright', legend = colnames(USArrests)[c(1, 4)], fill = 1:2)
}


set.seed(1)
tmp <- rnorm(25)
qqnorm(tmp)
qqline(tmp)
pars <- subplot(hist(tmp, ann = FALSE), 0, -1)

## wrong way to add a reference line to histogram
abline(v = 0, col = 2, lwd = 2)

## right way to add a reference line to histogram
op <- par(no.readonly = TRUE)
par(pars)
abline(v = 0, col = 3, lwd = 2)
par(op)

}
\seealso{
\code{TeachingDemos::subplot}; \url{http://sickel.net/blogg/?p=688};
\code{\link{inset}}
}
