% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_extra.R
\name{subrect}
\alias{subrect}
\title{subrect}
\usage{
subrect(
  xleft,
  ybottom,
  xright,
  ytop,
  type = c("diagonal", "square"),
  pos = c("topleft", "topright", "bottomleft", "bottomright"),
  ...
)
}
\arguments{
\item{xleft, ybottom, xright, ytop}{coordinates of the main \code{rect}}

\item{type, pos}{the type (diagonal or square) and position of the shape
inside the \code{rect}}

\item{...}{additional arguments passed to \code{\link{rect}}}
}
\description{
Draw shapes inside a \code{\link{rect}}
}
\examples{
pos <- eval(formals(subrect)$pos)
par(mfrow = n2mfrow(length(pos)))

for (p in pos) {
  plot(0, 0, main = p, type = 'n', xlab = '', ylab = '')
  rect(-0.5, -0.5, 0.5, 0.5)
  subrect(-0.5, -0.5, 0.5, 0.5, type = 'diag', pos = p, col = 'red')
}

for (p in pos) {
  plot(0, 0, main = p, type = 'n', xlab = '', ylab = '')
  rect(-0.5, -0.5, 0.5, 0.5)
  subrect(-0.5, -0.5, 0.5, 0.5, type = 'square', pos = p, col = 'red')
}

}
