% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2.R
\name{tableplot}
\alias{tableplot}
\title{tableplot}
\usage{
tableplot(
  x,
  y = NULL,
  table,
  title = NULL,
  rownames = NULL,
  colnames = NULL,
  bg = "transparent",
  cex = par("cex"),
  cex.title = cex,
  xjust = 0,
  yjust = 1,
  xpad = 0.25,
  ypad = 0.75,
  col.table = 1L,
  col.title = col.table,
  col.colnames = col.table,
  col.rownames = col.table,
  font.table = 1L,
  font.title = font.table,
  font.colnames = font.table,
  font.rownames = font.table,
  show.colnames = TRUE,
  show.rownames = FALSE,
  hlines = FALSE,
  vlines = FALSE,
  frame.colnames = FALSE,
  frame.title = FALSE,
  frame.table = FALSE,
  frame.rownames = FALSE,
  frame.type = c("box", "box", "line", "box"),
  ...
)
}
\arguments{
\item{x, y}{the x- and y-coordinates to be used or a keyword; see
\code{\link{xy.coords}}}

\item{table}{a data frame or matrix}

\item{title}{optional title for the table}

\item{rownames, colnames}{optional labels for rows and columns (default
are \code{rownames(table)} and \code{colnames(table)}, respectively)}

\item{bg}{background color for each table cell, recycled as needed}

\item{cex, cex.title}{text size for table, title}

\item{xjust, yjust}{table position relative to \code{x} and \code{y}}

\item{xpad, ypad}{padding around text in cells as a proportion of the max
width and height of the strings in each column}

\item{col.table, col.title, col.colnames, col.rownames}{colors for table,
title, column names, and row names}

\item{font.table, font.title, font.colnames, font.rownames}{fonts for table,
title, column names, and row names}

\item{show.colnames, show.rownames}{logical; if \code{TRUE}, column and row
names of \code{table} are added}

\item{hlines, vlines}{logical; if \code{TRUE}, horizontal and/or vertical
lines are drawn between table cells}

\item{frame.colnames, frame.title, frame.table, frame.rownames}{logical; if
\code{TRUE}, draws a frame around the object}

\item{frame.type}{a vector giving the type of framing for column names,
title, table, and row names, respectively; options are \code{"box"} or
\code{"line"}; alternatively, a character string with the first letter
for each, e.g., \code{"blbl"}}

\item{...}{additional graphical parameters passed to \code{\link{par}}}
}
\value{
A list with components \code{table}, \code{colnames}, \code{title}, and
\code{outline} giving the coordinates of the rectangle for each.
}
\description{
Add a table to a base plot.
}
\examples{
tbl <- head(mtcars, 3)
plot(mpg ~ wt, mtcars)

tableplot(
  'topright', table = tbl,
  title = 'mtcars data set', cex.title = 2
)

tableplot(
  par('usr')[1], 35, tbl[, 1:3],
  show.rownames = TRUE, col.rownames = 1:3,
  font.colnames = 2, hlines = TRUE
)

tableplot(
  'bottomleft', table = tbl[, 1:3], col.table = 1:3,
  show.rownames = TRUE, frame.rownames = TRUE, frame.type = 'bbll'
)

}
\seealso{
\code{\link[plotrix]{addtable2plot}}
}
