% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2.R
\name{tallbar}
\alias{tallbar}
\title{Tall barplots}
\usage{
tallbar(height, labels = names(height), groups = NULL, horiz = TRUE, ...)
}
\arguments{
\item{height}{a list of \code{\link{barplot}}s having either a single row
or a common row variable}

\item{labels, groups}{labels for each bar plot in \code{height} and
(optionally) for groups of bar plots}

\item{horiz}{logical the direction of the bars}

\item{...}{additional arguments passed to \code{\link{barplot2}}}
}
\value{
The default return of \code{\link{barplot2}} plus additional elements
\code{names}, \code{labels}, and \code{groups} giving the tick coordinates
and labels for each, useful for finer control over the axes; see examples.
}
\description{
Stack multiple bar plots vertically (or horizontally).
}
\examples{
x <- mtcars[, sapply(mtcars, function(x) length(unique(x)) < 5)]
y <- sapply(x, simplify = FALSE, function(x) t(table(x, mtcars$mpg > 25)))
x <- sapply(x, simplify = FALSE, function(x) t(table(x)))

tallbar(x, horiz = FALSE)
tallbar(x, las = 1)
tallbar(x, las = 1, groups = c('Group 2' = 2, 'Group 3' = 4))

bp <- tallbar(
  y, horiz = TRUE, las = 1, axisnames = FALSE,
  groups = c('Group 2' = 2, 'Group 3' = 4)
)
axis(2, bp$names,  names(bp$names),  las = 1, lwd = 0)
axis(2, bp$labels, names(bp$labels), las = 1, lwd = 0, font = 3)
axis(2, bp$groups, names(bp$groups), las = 1, lwd = 0, font = 2)

}
