% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_border}
\alias{theme_border}
\title{Theme with borders}
\usage{
theme_border()
}
\value{
A list of ggplot2 theme elements
}
\description{
A ggplot2 theme with clean borders
}
\examples{
\dontrun{
library(ggplot2)
ggplot(mtcars, aes(wt, mpg)) + 
  geom_point() + 
  theme_border()
}
}
