% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_gridlines}
\alias{theme_gridlines}
\title{Theme with customized grid lines}
\usage{
theme_gridlines()
}
\value{
A theme object that can be added to a ggplot object
}
\description{
A ggplot2 theme with lighter grid lines
}
\examples{
\dontrun{
library(ggplot2)
ggplot(mtcars, aes(wt, mpg)) + 
  geom_point() + 
  theme_gridlines()
}
}
