% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_line}
\alias{theme_line}
\title{Theme with axis lines}
\usage{
theme_line()
}
\value{
A list of ggplot2 theme elements
}
\description{
A ggplot2 theme with axis lines instead of full borders
}
\examples{
\dontrun{
library(ggplot2)
ggplot(mtcars, aes(wt, mpg)) + 
  geom_point() + 
  theme_line()
}
}
