% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_noaxisticks}
\alias{theme_noaxisticks}
\title{Theme without axis ticks}
\usage{
theme_noaxisticks()
}
\value{
A theme object that can be added to a ggplot object
}
\description{
A ggplot2 theme with no axis ticks or grid lines
}
\examples{
\dontrun{
library(ggplot2)
ggplot(mtcars, aes(wt, mpg)) + 
  geom_point() + 
  theme_noaxisticks()
}
}
