% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_scatter}
\alias{theme_scatter}
\title{Theme for scatter plots}
\usage{
theme_scatter()
}
\value{
A theme object that can be added to a ggplot object
}
\description{
A ggplot2 theme with no grid lines and axis text
}
\examples{
\dontrun{
library(ggplot2)
ggplot(mtcars, aes(wt, mpg)) + 
  geom_point() + 
  theme_scatter()
}
}
