% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_text}
\alias{theme_text}
\title{Theme for text elements}
\usage{
theme_text(
  base_family = "Helvetica",
  base_size = 14,
  title_size = 16,
  title_face = "bold",
  legend_title_size = 14,
  legend_text_size = 12,
  axis_title_size = 14,
  axis_text_size = 14,
  x_text_angle = 0,
  x_text_hjust = 0.5
)
}
\arguments{
\item{base_family}{Base font family. Default is "Helvetica".}

\item{base_size}{Base font size. Default is 14.}

\item{title_size}{Plot title size. Default is 16.}

\item{title_face}{Plot title face. Default is "bold".}

\item{legend_title_size}{Legend title size. Default is 14.}

\item{legend_text_size}{Legend text size. Default is 12.}

\item{axis_title_size}{Axis title size. Default is 14.}

\item{axis_text_size}{Axis text size. Default is 14.}

\item{x_text_angle}{X axis text angle. Default is 0.}

\item{x_text_hjust}{X axis text horizontal justification. Default is 0.5.}
}
\value{
A theme object that can be added to a ggplot object
}
\description{
Apply consistent text styling to ggplot2 graphics
}
\examples{
\dontrun{
library(ggplot2)
ggplot(mtcars, aes(wt, mpg)) + 
  geom_point() + 
  theme_text()
}
}
