% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_functions.R
\name{title_ggplots}
\alias{title_ggplots}
\title{Save plots to a PDF file}
\usage{
title_ggplots(plotlist, titles)
}
\arguments{
\item{plotlist}{A list of ggplot objects}

\item{titles}{A vector of titles for the plots}
}
\value{
A ggplot object
}
\description{
Save plots to a PDF file
}
