% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2.R
\name{toxplot}
\alias{toxplot}
\title{toxplot}
\usage{
toxplot(
  ftable,
  total,
  digits = 0L,
  headers = NULL,
  xlim = NULL,
  order = c("none", "decreasing", "increasing"),
  label.bars = c("n", "percent", "none"),
  col.bars = NULL,
  digits.bars = digits,
  widths = c(1, 1, 2),
  space = 1,
  xaxis.at = NULL,
  col.bg = "grey",
  alpha.bg = c(0.25, 0.5),
  legend = TRUE,
  args.legend = list(),
  ...
)
}
\arguments{
\item{ftable}{a two- or three-dimensional \code{\link{ftable}}}

\item{total}{total number of patients (for percentages)}

\item{digits}{number of places past the decimal to keep for percentages}

\item{headers}{a vector of labels for text column headers}

\item{xlim}{the x-axis limits for the bar plot}

\item{order}{optional ordering for the rows, one of \code{"none"} (default),
\code{"decreasing"}, or \code{"increasing"}; note that 3-d tables should
not be re-ordered}

\item{label.bars}{optional text on each bar, one of \code{"n"} (default),
\code{"percent"}, or \code{"none"}}

\item{col.bars}{a vector of colors for each column of \code{ftable}}

\item{digits.bars}{number of places past the decimal to keep for
\code{label.bars} text}

\item{widths}{the relative width of each column}

\item{space}{additional space between final text column and bar plot}

\item{xaxis.at}{x-axis tick positions}

\item{col.bg, alpha.bg}{background color for text columns with alternating
\code{alpha.bg} transparency}

\item{legend}{logical; if \code{TRUE}, a legend for the \code{col.bars} is
shown}

\item{args.legend}{an optional \emph{named} list of arguments passed to
\code{\link{legend}}}

\item{...}{additional graphical parameters passed to \code{\link{par}}}
}
\description{
Draw a toxicity summary plot.
}
\examples{
## example from rawr::tabler_by
set.seed(1)

f <- function(x, ...) sample(x, 100, replace = TRUE, ...)
tox <- data.frame(
  id = rep(1:10, 10), phase = 1:2,
  code = f(rawr::ctcae_v4$tox_code[1:100]),
  grade = f(1:3, prob = c(.6, .3, .1)),
  stringsAsFactors = FALSE
)
tox <- cbind(tox, rawr::match_ctc(tox$code)[, c('tox_cat', 'tox_desc')])

t1 <- ftable(
  Category = tox$tox_cat,
  Description = tox$tox_desc,
  Grade = tox$grade
)
t2 <- ftable(
  Description = tox$tox_desc,
  Grade = tox$grade
)
n <- 25


## basic usage
toxplot(t1, n) ## three column
toxplot(t2, n) ## two column


toxplot(
  t2, n, widths = c(1, 3), label.bars = 'percent',
  headers = c('Toxicity description', 'N (\%)'),
  order = 'dec', xlim = c(0, 10), xaxis.at = c(0, 5, 10),
  args.legend = list(x = 'bottomright', horiz = FALSE, cex = 2)
)

}
