% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2.R
\name{tracebar}
\alias{tracebar}
\title{Tracing barplot}
\usage{
tracebar(height, col = NULL, pad = 0.05, alpha = 0.75, ...)
}
\arguments{
\item{height}{a matrix of data describing the bars which make up the plot,
usually a main/total bar and one or more smaller or subsets of the total;
bars are grouped by columns; the first row will be the main bar and other
rows will be minor bars}

\item{col}{colors for each row of \code{height}}

\item{pad}{numeric vector of length one or two controlling the horizontal
and vertical padding, respectively}

\item{alpha}{numeric value in \code{[0,1]} for the alpha transparency of
polygons}

\item{...}{additional arguments passed to \code{\link{barplot}}}
}
\value{
A numeric vector with the midpoint of each main bar (i.e., identical to a
call to \code{barplot(..., beside = FALSE)}).

Additionally, a list attribute (\code{attr(., "coords")}) where each list
element corresponds to a group of bars. The matrices contain coordinates
of each rectangle plus midpoints.
}
\description{
Draw a \code{\link{barplot}} with polygons tracing each row.
}
\examples{
set.seed(1)
tbl <- sapply(1:3, function(x) sort(rpois(3, 10), decreasing = TRUE))
tracebar(tbl)

tracebar(replace(tbl, 5, 0), col = 1:3, space = 0.5)

}
\seealso{
\code{\link{inbar}}
}
