% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trellis_png.R
\name{trellis_png}
\alias{trellis_png}
\title{Save Trellis Plot to PNG}
\usage{
trellis_png(x, file, prt = TRUE, ...)
}
\arguments{
\item{x}{a \code{trellis} object or list of \code{trellis} objects
(see \code{trellis.object} from the \pkg{lattice} package).}

\item{file}{name of output file without extension, up to 500 characters.}

\item{prt}{\code{TRUE} to also print plot.}

\item{...}{more parameters passed to \code{png()}.}
}
\value{
Returns the file name of the output PNG image.
}
\description{
Saves one or more trellis plots as PNG files,
and optionally prints the plot(s).
}
\examples{
# directory for PNG files
dir.create("plot", showWarnings = FALSE, recursive = TRUE)

# trellis plots
tpl <- lattice::xyplot(sin(1:100) ~ cos(1:100), type = "l")
tpp <- lattice::xyplot(sin(1:100) ~ cos(1:100))

trellis_png(tpl, file = "plot/sincos", width = 7, height = 7)
trellis_png(list(tpl, tpp), file = "plot/sincos", width = 7, height = 7)
}
\author{
Nathan Lucas
}
