% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{tsne}
\alias{tsne}
\title{t-SNE plot}
\usage{
tsne(data, group = rep(1L, ncol(data)), col = NULL, ..., perplexity = 30,
  centers = NULL, plot3d = FALSE, labels = FALSE,
  legend = length(perplexity) == 1L, args.legend = list())
}
\arguments{
\item{data}{an \code{r x c} matrix-like object with \code{r} features and
\code{c} samples}

\item{group}{a length \code{c} vector identifying the group of each sample}

\item{col}{a vector of colors for each group}

\item{...}{additional arguments passed to \code{\link{plot}} or further to
\code{\link{par}}}

\item{perplexity}{perplexity parameter passed to \code{\link[Rtsne]{Rtsne}}}

\item{centers}{an integer value for the number of desired clusters, passed
to \code{\link[stats]{kmeans}}}

\item{plot3d}{logical; if \code{TRUE}, the first three dimensions are drawn
as a 3D scatterplot; see \code{\link[rgl]{plot3d}}}

\item{labels}{argument controlling labels for points; if \code{FALSE}, no
labels are drawn; if a numeric value in \code{(0,1)}, points outside
quantile are labelled; otherwise, a vector of labels for each point}

\item{legend}{logical; if \code{TRUE}, a legend will be drawn}

\item{args.legend}{a \emph{named} list of \code{\link{legend}} arguments to
override defaults}
}
\description{
Wrapper function to plot an \code{\link[Rtsne]{Rtsne}} object with optional
\code{\link[stats]{kmeans}} groupings. Alternatively, plot a 3D scatterplot
of the first three t-SNE dimensions.
}
\examples{
## feature x sample matrix
dat <- t(unique(iris[, 1:4]))
grp <- unique(iris)$Species

tsne(dat)
tsne(dat, grp)
tsne(dat, grp, centers = 3)$km$perplexity30$cluster
tsne(dat, grp, centers = 3, perplexity = c(5, 10, 20, 30), pch = 16)
tsne(dat, grp, centers = 3, perplexity = 20, col = tcol(3:5, alpha = 0.5),
     pch = 16, cex = 1.5, las = 1L, bty = 'l', labels = 0.001,
     args.legend = list(x = 'topleft', title = 'Species'))

}
\seealso{
\code{\link{pca}}
}
