% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{umap_aes}
\alias{umap_aes}
\title{Theme for dimension reduction plots}
\usage{
umap_aes(type = "UMAP", aspect_ratio = 1, border_size = 0.8)
}
\arguments{
\item{type}{Reduction type - "UMAP", "tSNE", or "PC". Default is "UMAP".}

\item{aspect_ratio}{Aspect ratio for the plot. Default is 1.}

\item{border_size}{Border size in mm. Default is 0.8.}
}
\value{
A list of ggplot2 theme elements
}
\description{
Special aesthetics for dimension reduction plots like UMAP, tSNE, or PCA
}
\examples{
\dontrun{
library(ggplot2)
# Assuming df has UMAP1 and UMAP2 columns
ggplot(df, aes(UMAP1, UMAP2, color = cluster)) + 
  geom_point() + 
  umap_aes()
}
}
