% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack.R
\name{unpack}
\alias{unpack}
\title{Assigns variables from package-level environments to function environments}
\usage{
unpack()
}
\arguments{
\item{var}{variable name to export}

\item{package.envir}{package-level environment}

\item{user.envir}{user-level environment}
}
\description{
Moves all parameters (loaded palette, plot settings and graphical parameters) to the parent environment.
}
\keyword{Workhorse}
