% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2.R
\name{upset}
\alias{upset}
\title{upset}
\usage{
upset(
  x,
  xat = NULL,
  yat = NULL,
  col = c("black", "white"),
  labels.x = NULL,
  labels.y = rownames(x),
  col.na = "grey75",
  col.rect = c("grey95", "transparent"),
  fullrange = FALSE
)
}
\arguments{
\item{x}{a numeric or integer matrix, coerced to binary with 0s and
\code{NA}s treated as false/missing, i.e., only non-zero and non-missing
values are used}

\item{xat, yat}{the x- and y-coordinates for the columns and rows of
\code{x}, respectively}

\item{col}{the colors for \code{x > 0} and 0s/\code{NA}s, respectively;
note that 0s/\code{NA}s only appear if \code{fullrange = TRUE}}

\item{labels.x, labels.y}{(optional) labels for the columns and rows}

\item{col.na, col.rect}{colors for background placeholder points and
bars for each row of \code{x}; default is alternating grey and white}

\item{fullrange}{logical; if \code{TRUE}, open circles will be shown for
any 0s or \code{NA}s in \code{x}}
}
\description{
Add an upset panel to an existing plot.
}
\examples{
x <- t(mtcars[, c('cyl', 'vs', 'am', 'gear')])
y <- t(!!mtcars[, c('cyl', 'vs', 'am')])

op <- par(mar = c(9, 5, 4, 1), las = 1L)

bp <- barplot(x, col = 2:5, axisnames = FALSE)
text(bp, -4, colnames(x), srt = 45, xpd = NA, adj = 1)
box(bty = 'u')

yat <- seq(1, 3, length.out = nrow(y))
upset(y, xat = bp, yat = -yat, col.na = NA, fullrange = TRUE)
upset(y, xat = bp, yat = par('usr')[4L] + yat, col.rect = adjustcolor(2:4, 0.5))

par(op)

}
