% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.envir.tools.R
\name{view.palette}
\alias{view.palette}
\title{View available color palettes}
\usage{
view.palette(palette)
}
\arguments{
\item{palette}{name of color palette. If empty, all palettes are shown.}
}
\description{
View palette presets available by printing them on the console.
 Colors can be assigned to a variable if `palette` is not empty.
}
\keyword{"Color}
\keyword{Palette"}
