% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plots.R
\name{widebars}
\alias{widebars}
\title{Wide bar plots}
\usage{
widebars(x, y, main, sub, foot, note, col = c("grey90", "grey30",
  "lightblue2", "dodgerblue2"))
}
\arguments{
\item{x}{vector of grouping values}

\item{y}{numeric vector of data values}

\item{main}{overall title for plot}

\item{sub}{sub-title for plot}

\item{foot}{footnote for plot}

\item{note}{note for plot}

\item{col}{a vector of colors for elements of the form
\code{c(section background, text, negative bars, positive bars)}}
}
\description{
Create a series of barplots by grouping factor.
}
\examples{
\dontrun{
pdf('./widebar.pdf', width = 14, height = 7)
par(oma = c(5,5,7,5), mar = c(5,0,3,.1), las = 1,
    fg = 'transparent', bg = 'grey98')
widebars(x <- rep(2004:2015, each = 4), rnorm(48),
         main = sprintf('Market value, \%s',
                        paste0(range(x), collapse = ' - ')),
         sub = 'Percent change, quarterly',
         foot = sprintf('Values current as of \%s',
                        format(Sys.time(), '\%b \%Y')),
         note = 'github.com/raredd')
dev.off()
}

par(oma = c(5,5,7,5), mar = c(5,0,3,.1), las = 1,
    fg = 'transparent', bg = 'grey98')
widebars(mtcars$gear, mtcars$mpg * sample(c(-1,1), 32, replace = TRUE),
         'Motor Trend car road tests', 'Miles per gallon by gear',
         'some footnote that isn\\'t important', 'blahblah')
}

