% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_extra.R
\name{zoomin}
\alias{zoomin}
\title{Zoom for points in base \code{R} plot}
\usage{
zoomin(x, y, ...)
}
\arguments{
\item{x, y}{x- and y-axis variables}

\item{...}{additional arguments passed to \code{\link{identify}}}
}
\description{
Provides a summary statistic for sample points in a plot.
}
\examples{
set.seed(1)
x <- runif(10)
y <- rnorm(10, mean = 5)

par(mfrow = c(1, 2))
plot(x, y, xlab = 'mean', ylab = 'sd')

zoomin(x, y)
## ESC to quit

}
