% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_special.R
\name{bump}
\alias{bump}
\title{bump chart}
\usage{
bump(mat, adj = 0.1, ...)
}
\arguments{
\item{mat}{an \code{n x t} matrix with \code{n} observations and \code{t}
timepoints}

\item{adj}{label position adjustment, larger values move labels farther
from start/end positions}

\item{...}{additional graphical parameters passed to \code{\link{par}}}
}
\description{
A bump chart
}
\examples{
mat <- replicate(5, sample(1:10))
dimnames(mat) <- list(rownames(mtcars)[1:nrow(mat)],
                      paste0('time', 1:ncol(mat)))
bump(mat, mar = c(2, 0, 2, 9))

}
