% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{dimr}
\alias{dimr}
\alias{plot.dimr}
\title{Dimension reduction}
\usage{
dimr(data, type = c("tsne", "pca", "rpca"), k = NULL, ...)

\method{plot}{dimr}(x, group = TRUE, group2 = NULL, k = NULL,
  col = NULL, labels = FALSE, vegan = FALSE, legend = TRUE,
  args.legend = list(), plot3d = FALSE, iplot = FALSE, ...)
}
\arguments{
\item{data}{an \code{r x c} matrix-like object with \code{r} features and
\code{c} samples}

\item{type}{the type of algorithm to run}

\item{k}{an integer representing the number of desired clusters, passed to
\code{\link[stats]{kmeans}}; note this is only used for convenience and
does not affect the algorithm determied by \code{type}

note, if \code{k} is passed in \code{plot.dimr}, a new k-means clustering
is performed; this is useful for on-the-fly clustering}

\item{...}{for \code{dimr}, additional arguments passed to the algorithm
determined by \code{type}; for \code{plot.dimr}, additional arguments
passed to plotting functions or further to \code{\link{par}}}

\item{x}{an object of class \code{"dimr"}}

\item{group}{a length \code{c} vector identifying the group of each sample;
use \code{FALSE} for none or \code{TRUE} to use k-means clusters}

\item{group2}{a list of data to plot, usually samples in \code{data}, each
to be shown as a scatter plot}

\item{col}{a vector of colors for each group}

\item{labels}{argument controlling labels for points; if \code{FALSE}, no
labels are drawn; if a numeric value in \code{(0,1)}, points outside
quantile are labelled; otherwise, a vector of labels for each point}

\item{vegan}{logical; if \code{TRUE}, ordination diagrams are drawn for each
k-means cluster; see \code{\link[vegan]{ordihull}}}

\item{legend}{logical; if \code{TRUE}, a legend will be drawn}

\item{args.legend}{a \emph{named} list of \code{\link{legend}} arguments to
override defaults}

\item{plot3d}{logical; if \code{TRUE}, the first three dimensions are drawn
as a 3D scatterplot; see \code{\link[rgl]{plot3d}}}

\item{iplot}{logical; if \code{TRUE}, the first two dimensions are drawn
as an interactive scatter plot; see \code{\link[iplotr]{iscatter}}}
}
\value{
\code{dimr} returns a list of class "\code{dimr}" with the following
components:

\item{\code{object}}{the object return by the algorithm determined via
\code{type}}
\item{\code{x}}{a matrix with the first three dimensions}
\item{\code{kmeans}}{a k-means clustering performed on the first two
dimensions of \code{object}}
}
\description{
Wrapper function to perform and plot methods of dimensionality reduction
including \code{\link[Rtsne]{Rtsne}}, \code{\link[rsvd]{rpca}}, and
\code{\link[stats]{prcomp}} with optional \code{\link[stats]{kmeans}}
clusters.
}
\examples{
## feature x sample matrix
dat <- t(unique(iris[, 1:4]))
grp <- unique(iris)$Species

tsne <- dimr(dat)
pca  <- dimr(dat, type = 'pca')
rpca <- dimr(dat, type = 'rpca')
pca2 <- dimr(dat, type = 'pca', scale. = TRUE) ## same as rpca

plot(tsne)        ## use k-means cluster groups
plot(tsne, grp)   ## use groups from data
plot(tsne, k = 3) ## perform new k-means

plot(pca, vegan = TRUE)
plot(pca, k = 4, vegan = TRUE)$kmeans$cluster


## use group2 to show scatter plots of numeric data
plot(pca, group2 = as.list(data.frame(t(dat))))
plot(pca, pch = 16, cex = 3, col = adjustcolor(palette()[1:3], alpha.f = 0.5),
     group2 = as.list(data.frame(t(rbind(dat, dat)))))


## use plot3d = TRUE for a 3D, interactive figure
\dontrun{
plot(tsne, plot3d = TRUE)

plot(tsne, iplot = TRUE, labels = data.frame(t(dat)))


## make a gridplot (called from plot.dimr)
n <- 9
d <- as.list(mtcars[, rep_len(c('mpg', 'wt', 'hp'), n)])
par(mfrow = n2mfrow(n), oma = c(5, 5, 4, 2))
plotr:::gridplot(d, mtcars$mpg, mtcars$wt, legend = TRUE)
title(xlab = 'MPG', ylab = 'Weight', outer = TRUE, cex.lab = 2)
}

}
\seealso{
\code{\link[Rtsne]{Rtsne}}; \code{\link[rsvd]{rpca}};
\code{\link[stats]{prcomp}}; \code{\link[stats]{kmeans}}
}
