% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplots.R
\name{gbxp}
\alias{gbxp}
\title{Draw boxplots from summaries}
\usage{
gbxp(z, notch = FALSE, width = NULL, varwidth = FALSE,
  outline = TRUE, notch.frac = 0.5, log = "", border = par("fg"),
  pars = NULL, horizontal = FALSE, add = FALSE, at = NULL,
  show.names = NULL, ..., grid = TRUE, col.grid = "grey90",
  col.acc = "white")
}
\arguments{
\item{z}{a list containing data summaries to be used in constructing
    the plots.  These are usually the result of a call to
    \code{\link{boxplot}}, but can be generated in any fashion.}

\item{notch}{if \code{notch} is \code{TRUE}, a notch is drawn in each
    side of the boxes.  If the notches of two plots do not overlap then
    the medians are significantly different at the 5 percent level.}

\item{width}{a vector giving the relative widths of the boxes making
    up the plot.}

\item{varwidth}{if \code{varwidth} is \code{TRUE}, the boxes are drawn
    with widths proportional to the square-roots of the number of
    observations in the groups.}

\item{outline}{if \code{outline} is not true, the outliers are not
    drawn.}

\item{notch.frac}{numeric in (0,1). When \code{notch = TRUE}, the
    fraction of the box width that the notches should use.}

\item{log}{character, indicating if any axis should be drawn in
    logarithmic scale, as in \code{\link{plot.default}}.}

\item{border}{character or numeric (vector), the color of the box
    borders.  Is recycled for multiple boxes.  Is used as default for
    the \code{boxcol}, \code{medcol}, \code{whiskcol}, \code{staplecol},
    and \code{outcol} options (see below).
  }

\item{pars}{\link{graphical parameters} (etc) can be passed as
    arguments to this function, either as a list (\code{pars}) or
    normally(\code{\dots}), see the following.  (Those in \code{\dots}
    take precedence over those in \code{pars}.)

    Currently, \code{yaxs} and \code{ylim} are used \sQuote{along the
      boxplot}, i.e., vertically, when \code{horizontal} is
    false, and \code{xlim} horizontally.
    \code{xaxt}, \code{yaxt}, \code{las}, \code{cex.axis}, and
    \code{col.axis} are passed to \code{\link{axis}}, and \code{main},
    \code{cex.main}, \code{col.main}, \code{sub}, \code{cex.sub},
    \code{col.sub}, \code{xlab}, \code{ylab}, \code{cex.lab}, and
    \code{col.lab} are passed to \code{\link{title}}.

    In addition, \code{axes} is accepted (see
    \code{\link{plot.window}}), with default \code{TRUE}.

    The following arguments (or \code{pars} components) allow further
    customization of the boxplot graphics.  Their defaults are typically
    determined from the non-prefixed version (e.g., \code{boxlty} from
    \code{lty}), either from the specified argument or \code{pars}
    component or the corresponding \code{\link{par}} one.
    \describe{
      \item{boxwex:}{a scale factor to be applied to all boxes.  When there
        are only a few groups, the appearance of the plot can be improved by
        making the boxes narrower.  The default depends on \code{at} and
        typically is \eqn{0.8}.}
      \item{staplewex, outwex:}{staple and outlier line width expansion,
        proportional to box width; both default to 0.5.}
      \item{boxlty, boxlwd, boxcol, boxfill:}{box outline type, width,
        color, and fill color (which currently defaults to \code{col} and will
        in future default to \code{par("bg")}).}
      \item{medlty, medlwd, medpch, medcex, medcol, medbg:}{median line type,
        line width, point character, point size expansion, color, and
        background color.  The default \code{medpch = NA} suppresses the
        point, and \code{medlty = "blank"} does so for the line.
        Note that\code{medlwd} defaults to
        \eqn{3\times}{3x} the default \code{lwd}.}
      \item{whisklty, whisklwd, whiskcol:}{whisker line type (default:
        \code{"dashed"}), width, and color.}
      \item{staplelty, staplelwd, staplecol:}{staple (= end of whisker) line
        type, width, and color.}
      \item{outlty, outlwd, outpch, outcex, outcol, outbg:}{outlier line
        type, line width, point character, point size expansion, color, and
        background color.  The default \code{outlty =  "blank"} suppresses the
        lines and \code{outpch = NA} suppresses points.}
    }
  }

\item{horizontal}{logical indicating if the  boxplots should be
    horizontal; default \code{FALSE} means vertical boxes.}

\item{add}{logical, if true \emph{add} boxplot to current plot.}

\item{at}{numeric vector giving the locations where the boxplots
    should be drawn, particularly when \code{add = TRUE}; defaults to
    \code{1:n} where \code{n} is the number of boxes.}

\item{show.names}{Set to \code{TRUE} or \code{FALSE} to override the
    defaults on whether an x-axis label is printed for each group.}

\item{grid}{logical; if \code{TRUE}, a background grid will be drawn}

\item{col.grid}{\code{grid} color}

\item{col.acc}{\code{grid} accent color}
}
\description{
\code{gbxp} draws box plots based on the given summaries in \code{z}. It is
usually called from within \code{\link{gboxplot}} but can be invoked 
directly.
}
