% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{get_palette}
\alias{get_palette}
\title{Get color palette}
\usage{
get_palette(palette, n = 9, direction = 1)
}
\arguments{
\item{palette}{A function that returns a color palette, a vector of hexadecimal color codes,
the name of a palette from "RColorBrewer" or "viridis", or one of the package's palette functions.}

\item{n}{Number of colors to return. Default is 9.}

\item{direction}{Direction of the palette vector (1 for forward, -1 for reverse).}
}
\value{
A vector of hexadecimal color codes.
}
\description{
Get a vector of hexadecimal color codes based on the specified palette.
}
\examples{
\dontrun{
# Using a predefined package palette function
colors <- get_palette(pal_kelly_20, n = 5)

# Using a palette directly  
colors <- get_palette(pal_kelly_20(), n = 5)

# Using a viridis palette name
colors <- get_palette("viridis", n = 10)

# Using a ColorBrewer palette name
colors <- get_palette("Blues", n = 7)

# Reverse direction
colors <- get_palette(pal_kelly_20, direction = -1)
}
}
