% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.color.R
\name{is.color}
\alias{is.color}
\title{Test color validity}
\usage{
is.color(color)
}
\arguments{
\item{color}{vector of colors to test}
}
\value{
TRUE if all elements can be translated into an RGB color, FALSE otherwise
}
\description{
Check if all values in vector are coercible to RGB.
}
\keyword{Utilities}
