% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-dataSettings.R
\name{modelSettings}
\alias{modelSettings}
\title{Server function for model settings}
\usage{
modelSettings(input, output, session, modelParameters)
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session}

\item{modelParameters}{(reactive) list of settings for input fields}
}
\description{
Backend for model settings module
}
