% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotr.R
\name{plotr}
\alias{plotr}
\title{Scatter and line plots}
\usage{
plotr(formula, x, y, plot.param)
}
\arguments{
\item{formula}{formula relating the x- and y-coordinates. Only y ~ x structure is accepted}

\item{x}{x-coordinates. Incompatible with formula and requires input y}

\item{y}{y-coordinates. Incompatible with formula and requires input x}

\item{plot.param}{variable of type scatter.param containing plot parameters}
}
\description{
Generates scatter and/or line plot using either vector or matrix as the y-coordinates and a vector as the x-coordinates.
}
