% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotr_utils}
\alias{plotr_utils}
\alias{oror}
\alias{\%||\%}
\alias{notin}
\alias{\%ni\%}
\alias{inside}
\alias{\%inside\%}
\title{plotr utils}
\usage{
x \%ni\% table

x \%inside\% interval

a \%||\% b
}
\arguments{
\item{x}{vector or \code{NULL}; the values to be matched}

\item{table}{vector or \code{NULL}; the values to be matched against}

\item{interval}{numeric vector of length two representing the interval}

\item{a, b}{raw, logical, "number-like" vectors or objects}
}
\description{
Some utilities for \code{plotr}.
}
\details{
\code{\%ni\%} is the negation of \code{\link[base]{\%in\%}}.

\code{\%inside\%} returns a logical vector indicating if \code{x} is inside
the \code{interval} (inclusive).

\code{\%||\%} is useful for a function, \code{f}, that may return a value
or \code{NULL}, but if \code{NULL} is the result of \code{f}, it is
desirable to return some other default value without errors.
}
\examples{
1:5 \%ni\% 3:5

c(0,4) \%inside\% c(0, 4)
-5:5 \%inside\% c(0, 5)
-5:5 \%inside\% c(5, 0)

# NULL || TRUE ## error
NULL \%||\% TRUE ## TRUE
}
\seealso{
\code{\link{\%in\%}}, \code{\link{||}}
}
