% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{prettypie}
\alias{prettypie}
\alias{prettypie2}
\title{prettypie}
\usage{
prettypie(dat, file, dev = "pdf", width = 15, height = 11, main, sub,
  note)

prettypie2(x, group = 1, labels = NA, col = NULL, radius = c(0.7, 1))
}
\arguments{
\item{dat}{data}

\item{file}{path to output file}

\item{dev}{device; default is \code{\link{pdf}}}

\item{width}{width of \code{dev}}

\item{height}{height of \code{dev}}

\item{main}{overall title for plot}

\item{sub}{sub-title for plot}

\item{note}{note for plot}

\item{x}{numeric vector representing the size for each slice}

\item{group}{vector identifying groups of slices (used by \code{col})}

\item{labels}{vector of labels for individual slices; if \code{NA}, no
labels will be drawn; if \code{NULL}, slices will be labeled sequentially}

\item{col}{colors for each group of slices}

\item{radius}{radius for inner and outer pie (usually in [0,1])}
}
\description{
A pie chart. \code{prettypie} is stupid, ignore it; use \code{prettypie2},
}
\examples{
\dontrun{
browsers <- source(system.file('source','browsers.r', package = 'plotr'))$value
browsers <- within(browsers, total <- ave(share, browser, FUN = sum))
browsers <- browsers[c('browser','version','share','total')]

prettypie(dat = browsers,
          file = './donuts.pdf', 
          main = 'Browser market share, April 2011',
          sub = 'stackoverflow.com:::maryam',
          note = '/questions/26748069/ggplot2-pie-and-donut-chart-on-same-plot')


disease <- source(system.file('source','disease.r', package = 'plotr'))$value
disease <- disease[c('Disease','Acosts60','Pcosts60','Patients60','Total60')]

prettypie(dat = disease,
          file = './cost.pdf', 
          main = 'Cost of getting sick',
          sub = 'Inside: Personal Costs, outside: Insurer Costs',
          note = 'visualization.geblogs.com/visualization/health_costs/')
}
## basic usage
prettypie2(mtcars$mpg)

with(mtcars, prettypie2(mpg, interaction(gear, cyl), rownames(mtcars)))

}
