% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{scattergram}
\alias{scattergram}
\title{Scattergram}
\usage{
scattergram(x, y, g = NULL, col = NULL, ..., xlim = NULL,
  ylim = NULL, axs = "i", main = NULL, legend = TRUE,
  args.legend = list())
}
\arguments{
\item{x, y, g}{the x-, y-, and group variables}

\item{col}{a vector of colors or integers for each unique value of \code{g}}

\item{...}{additional arguments passed to \code{plot} (e.g., \code{xlim},
\code{panel.first}) or further to \code{par}}

\item{xlim, ylim}{the x- and y-axis limits}

\item{axs}{the style(s) of axis interval calculation; see \code{\link{par}}}

\item{main}{the main figure title}

\item{legend}{logical; if \code{TRUE}, a \code{group} legend is drawn}

\item{args.legend}{a \emph{named} list of \code{\link{legend}} arguments to
override defaults}
}
\description{
A scattergram plot with density curves along axes.
}
\examples{
with(mtcars, {
  scattergram(mpg, wt)
})

with(mtcars, {
  scattergram(mpg, wt, cyl, col = rainbow(3), pch = 16)
})

with(mtcars, {
  scattergram(
    mpg, wt, c('Four', 'Six', 'Eight')[cyl / 2 - 1],
    xlab = 'MPG', ylab = 'Weight', las = 1,
    panel.first = abline(lm(wt ~ mpg), col = 5, lwd = 3),
    panel.last = abline(v = c(20, 25))
  )
})

}
