% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-NAMESPACE.R, R/02-addMoreData.R,
%   R/02-saveAndDeletePlot.R, R/03-addMorePoints.R, R/03-downUploads.R,
%   R/03-goodnessFit.R, R/03-multiplePlots.R, R/03-multiplePredictions.R,
%   R/03-postProcessing.R, R/03-runModel.R, R/03-stylePlot.R, R/03-uploadFiles.R
\name{shinyModule}
\alias{shinyModule}
\alias{addMoreDataUI}
\alias{addMoreData}
\alias{savePlotUI}
\alias{savePlot}
\alias{deletePlotUI}
\alias{deletePlot}
\alias{addMorePointsUI}
\alias{addMorePoints}
\alias{downUploadsUI}
\alias{downUploads}
\alias{goodnessOfFitUI}
\alias{goodnessOfFit}
\alias{multiplePlotsUI}
\alias{multiplePlots}
\alias{multiplePredictionsUI}
\alias{multiplePredictions}
\alias{postProcessingUI}
\alias{postProcessing}
\alias{runModelUI}
\alias{runModel}
\alias{stylePlotUI}
\alias{stylePlot}
\alias{uploadFilesUI}
\alias{uploadFiles}
\title{Server and UI Functions for Shiny Module}
\usage{
addMoreDataUI(id, label)

addMoreData(input, output, session)

savePlotUI(id, label)

savePlot(input, output, session, savedData, currentPlot)

deletePlotUI(id, label)

deletePlot(input, output, session, savedData)

addMorePointsUI(id, title)

addMorePoints(input, output, session, savedData)

downUploadsUI(id, title)

downUploads(input, output, session, savedData, loadedFiles)

goodnessOfFitUI(id, title)

goodnessOfFit(input, output, session, savedData)

multiplePlotsUI(id, title)

multiplePlots(input, output, session, savedData)

multiplePredictionsUI(id, title)

multiplePredictions(input, output, session, savedData, loadedFiles)

postProcessingUI(id, title)

postProcessing(input, output, session, savedData)

runModelUI(id, title)

runModel(input, output, session, loadedFiles)

stylePlotUI(id, title)

stylePlot(input, output, session, savedData)

uploadFilesUI(id, title)

uploadFiles(input, output, session)
}
\arguments{
\item{id}{namespace id}

\item{label}{(character) label used in UI}

\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session}

\item{savedData}{(reactive) list of saved data}

\item{currentPlot}{plot object to be saved}

\item{title}{title of tab in tabset panel}

\item{loadedFiles}{(reactive) list of uploaded files}
}
\description{
Server and UI Functions for Shiny Module
}
