% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2.R
\name{tableplot}
\alias{tableplot}
\title{tableplot}
\usage{
tableplot(x, y = NULL, table, title = NULL, bg = "transparent",
  cex = 1, cex.title = cex, xjust = 0, yjust = 1, xpad = 0.25,
  ypad = 0.75, col.table = 1L, col.title = col.table,
  col.colnames = col.table, col.rownames = col.table,
  font.table = 1L, font.title = font.table,
  font.colnames = font.table, font.rownames = font.table,
  show.colnames = TRUE, show.rownames = FALSE, hlines = FALSE,
  vlines = FALSE, frame.colnames = FALSE, frame.title = FALSE,
  frame.table = FALSE, frame.type = c("box", "box", "line"))
}
\arguments{
\item{x, y}{the x- and y-coordinates to be used or a keyword; see
\code{\link{xy.coords}}}

\item{table}{a data frame or matrix}

\item{title}{optional title for the table}

\item{bg}{background color for each table cell, recycled as needed}

\item{cex, cex.title}{text size for table, title}

\item{xjust, yjust}{table position relative to \code{x} and \code{y}}

\item{xpad, ypad}{padding around text in cells as a proportion of the max
width and height of the strings in each column}

\item{col.table, col.title, col.colnames, col.rownames}{colors for table,
title, column names, and row names}

\item{font.table, font.title, font.colnames, font.rownames}{fonts for table,
title, column names, and row names}

\item{show.colnames, show.rownames}{logical; if \code{TRUE}, column and row
names of \code{table} are added}

\item{hlines, vlines}{logical; if \code{TRUE}, horizontal and/or vertical
lines are drawn between table cells}

\item{frame.colnames, frame.title, frame.table}{logical; if \code{TRUE},
draws a frame around the object}

\item{frame.type}{the type of framing for column names, title, and table,
respectively; options are \code{"box"} or \code{"line"}}
}
\value{
A list with components \code{table}, \code{colnames}, \code{title}, and
\code{outline} giving the coordinates of the rectangle for each.
}
\description{
Add a table to a base plot.
}
\examples{
plot(mpg ~ wt, mtcars)
tableplot(
  'topright', table = head(mtcars, 3),
  title = 'mtcars data set', cex.title = 2
)
tableplot(
  par('usr')[1], 35, head(mtcars, 3)[, 1:3],
  show.rownames = TRUE, col.rownames = 'red',
  font.colnames = 2, hlines = TRUE
)

}
\seealso{
\code{\link[plotrix]{addtable2plot}}
}
