% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2.R
\name{toxplot}
\alias{toxplot}
\title{toxplot}
\usage{
toxplot(data, headers = NULL, widths = c(1, 1, 1, 2),
  total = max(rowSums(tbl)), show.n = TRUE, col.bars = NULL,
  col.bg = "grey", alpha.bg = c(0.25, 0.5), digits = 0L,
  xaxis.at = 0:5/5, legend = TRUE, args.legend = list())
}
\arguments{
\item{data}{an \code{\link{ftable}} or a data frame with columns for toxicity
category, description, and grade (in that order)}

\item{headers}{a vector of labels for text column headers}

\item{widths}{the relative width of each column}

\item{total}{total number of patients (for percentages)}

\item{show.n}{logical; if \code{TRUE}, the count for each bar is shown}

\item{col.bars}{a vector of colors for each column of \code{ftable}}

\item{col.bg, alpha.bg}{background color for text columns with alternating
\code{alpha.bg} transparency}

\item{digits}{number of places past the decimal to keep for percentages}

\item{xaxis.at}{x-axis tick positions, scaled to \code{[0,1]}}

\item{legend}{logical; if \code{TRUE}, a legend for the \code{col.bars} is
shown}

\item{args.legend}{an optional \emph{named} list of arguments passed to
\code{\link{legend}}}
}
\description{
Draw a toxicity summary plot.
}
\examples{
## example from rawr::tabler_by
set.seed(1)

f <- function(x, ...) sample(x, 100, replace = TRUE, ...)
tox <- data.frame(
  id = rep(1:10, 10), phase = 1:2,
  code = f(rawr::ctcae_v4$tox_code[1:100]),
  grade = f(1:3, prob = c(.6, .3, .1)),
  stringsAsFactors = FALSE
)
tox <- cbind(tox, rawr::match_ctc(tox$code)[, c('tox_cat', 'tox_desc')])

tbl <- ftable(
  Category = tox$tox_cat,
  Description = tox$tox_desc,
  Grade = tox$grade
)

## these are equivalent ways to call toxplot
toxplot(tbl)
toxplot(tox[, c('tox_cat', 'tox_desc', 'grade')])

\dontrun{
pdf('~/desktop/test.pdf', height = 10, width = 12)
toxplot(tbl, xaxis.at = 0:1, args.legend = list(horiz = FALSE, cex = 2),
        col.bg = 'orchid', total = 15, widths = c(1, 1, 0.5, 2))
dev.off()
}

}
