% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{widebars}
\alias{widebars}
\title{Wide bar plots}
\usage{
widebars(x, y, main, sub, foot, note, col = c("grey90", "grey30",
  "lightblue2", "dodgerblue2"))
}
\arguments{
\item{x}{vector of grouping values}

\item{y}{numeric vector of data values}

\item{main}{overall title for plot}

\item{sub}{sub-title for plot}

\item{foot}{footnote for plot}

\item{note}{note for plot}

\item{col}{a vector of colors for elements of the form 
\code{c(section background, text, negative bars, positive bars)}}
}
\description{
Create a series of barplots by grouping factor.
}

